% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNfit-S3methods.R
\name{AIC.PLNfit}
\alias{AIC.PLNfit}
\title{Akaike Information Criterion for a fitted PLN model}
\usage{
\method{AIC}{PLNfit}(object, ..., k = 2)
}
\arguments{
\item{object}{an R6 object with class \code{\link{PLNfit}}}

\item{...}{additional parameters for S3 compatibility. Not used}

\item{k}{not used, present for S3 compatibility.}
}
\value{
A scalar: the variational AIC (larger is better).
}
\description{
Computes the variational AIC as \code{loglik - nb_param} (larger is better).
This follows the maximization convention used throughout PLNmodels.
}
\examples{
data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
model <- PLN(Abundance ~ 1, data = trichoptera)
AIC(model)
}
