% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barents.R
\docType{data}
\name{barents}
\alias{barents}
\title{Barents fish data set}
\format{
A data frame with 6 variables:
\itemize{
\item Abundance: A 30 fish species by 89 sites count matrix
\item Offset: A 30 fish species by 116 samples offset matrix, measuring the sampling effort in each site
\item 4 covariates for latitude, longitude, depth (in meters), temperature (in Celsius degrees).
}
}
\source{
Data from M. Fossheim and coauthors.
}
\usage{
barents
}
\description{
This data set gives the abundance of 30 fish species observed in 89 sites in the Barents sea.
For each site, 4 additional covariates are known. Subsample of the original datasets studied by Fossheim et al, 2006.
}
\examples{
data(barents)
}
\references{
Fossheim, Maria, Einar M. Nilssen, and Michaela Aschan. "Fish assemblages in the Barents Sea." Marine Biology Research 2.4 (2006). \doi{10.1080/17451000600815698}
}
\keyword{datasets}
