% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNfit-S3methods.R
\name{BIC.PLNfit}
\alias{BIC.PLNfit}
\title{Bayesian Information Criterion for a fitted PLN model}
\usage{
\method{BIC}{PLNfit}(object, ...)
}
\arguments{
\item{object}{an R6 object with class \code{\link{PLNfit}}}

\item{...}{additional parameters for S3 compatibility. Not used}
}
\value{
A scalar: the variational BIC (larger is better).
}
\description{
Computes the variational BIC as \code{loglik - 0.5 * log(n) * nb_param} (larger is better).
This follows the maximization convention used throughout PLNmodels.
}
\examples{
data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
model <- PLN(Abundance ~ 1, data = trichoptera)
BIC(model)
}
