% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNARfit-S3methods.R
\name{coef.PLNARfit}
\alias{coef.PLNARfit}
\title{Extract model coefficients}
\usage{
\method{coef}{PLNARfit}(object, type = c("main", "covariance"), ...)
}
\arguments{
\item{object}{an R6 object with class \code{\link{PLNARfit}}}

\item{type}{type of parameter that should be extracted. Either "main" (default) for \deqn{B} or "covariance" for \deqn{\Sigma}}

\item{...}{additional parameters for S3 compatibility. Not used}
}
\value{
A matrix of coefficients extracted from the PLNARfit model.
}
\description{
Extracts model coefficients from objects returned by \code{\link[=PLN]{PLN()}} and its variants
}
\examples{
data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPLN <- PLN(Abundance ~ 1 + offset(log(Offset)), data = trichoptera)
coef(myPLN) ## B
coef(myPLN, type = "covariance") ## Sigma
}
\seealso{
\code{\link[=sigma.PLNARfit]{sigma.PLNARfit()}}, \code{\link[=vcov.PLNARfit]{vcov.PLNARfit()}}, \code{\link[=standard_error.PLNARfit]{standard_error.PLNARfit()}}
}
