% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNfit-S3methods.R
\name{coef.PLNfit}
\alias{coef.PLNfit}
\title{Extracts model coefficients from objects returned by \code{\link[=PLN]{PLN}} and its variants}
\usage{
\method{coef}{PLNfit}(object, ...)
}
\arguments{
\item{object}{an R6 object with class PLNfit}

\item{...}{additional parameters for S3 compatibility. Not used}
}
\value{
A matrix of coefficients extracted from the PLNfit model.
}
\description{
Extracts model coefficients from objects returned by \code{\link[=PLN]{PLN}} and its variants
}
