% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNnetworkfamily-class.R
\name{coefficient_path}
\alias{coefficient_path}
\title{Extract the regularization path of a PLNnetwork fit}
\arguments{
\item{precision}{a logical, should the coefficients of the precision matrix Omega or the covariance matrice Sigma be sent back. Default is \code{TRUE}.}

\item{corr}{a logical, should the correlation (partial in case  precision = TRUE) be sent back. Default is \code{TRUE}.}
}
\value{
Send back a tibble/data.frame.
}
\description{
Extract the regularization path of a PLNnetwork fit
}
