% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_utils.R
\name{compute_offset}
\alias{compute_offset}
\title{Compute offsets from a count data using one of several normalization schemes}
\usage{
compute_offset(counts, offset = c("TSS", "GMPR", "RLE", "CSS", "none"),
  ...)
}
\arguments{
\item{counts}{Required. An abundance count table, preferably with dimensions names and species as columns.}

\item{offset}{Optional. Normalisation scheme used to compute scaling factors used as offset during PLN inference. Can be "TSS" (Total Sum Scaling, default), "CSS" (Cumulative Sum Scaling, used in metagenomeSeq), "RLE" (Relative Log Expression, used in DESeq2), "GMPR" (Geometric Mean of Pairwise Ratio, introduced in Chen et al., 2018) or "none".}

\item{...}{Additional parameters passed on to specific methods}
}
\value{
If offset = "none", NULL else a vector of length \code{nrow(counts)} with one offset per sample.
}
\description{
Computes offsets from the count table using one of several normalization schemes (TSS, CSS, RLE, GMPR, etc) described in the literature.
}
\details{
The only normalisation with an additional argument is RLE, for which you can add pseudocounts to the observed counts. The default is \code{pseudocounts = 0}. Note that (i) CSS normalization fails when the median absolute deviation around quantiles does not become instable for high quantiles, (ii) RLE fails when there are no common species across all samples and (iii) GMPR fails if a sample does not share any species with all other samples.
}
\examples{
data(trichoptera)
counts <- trichoptera$Abundance
compute_offset(counts)
}
\references{
Chen, L., Reeve, J., Zhang, L., Huang, S., Wang, X. and Chen, J. (2018) GMPR: A robust normalization method for zero-inflated count data with application to microbiome sequencing data. PeerJ, 6, e4600 \url{https://doi.org/10.7717/peerj.4600}

Paulson, J. N., Colin Stine, O., Bravo, H. C. and Pop, M. (2013) Differential abundance analysis for microbial marker-gene surveys. Nature Methods, 10, 1200-1202 \url{http://dx.doi.org/10.1038/nmeth.2658}

Anders, S. and Huber, W. (2010) Differential expression analysis for sequence count data. Genome Biology, 11, R106 \url{https://doi.org/10.1186/gb-2010-11-10-r106}
}
