% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNfit-S3methods.R
\name{fisher}
\alias{fisher}
\title{Generic function for the Fisher information matrix of Theta}
\usage{
fisher(x, ...)
}
\arguments{
\item{x}{An `R` object. Currently there are methods for \code{\link{PLNfit}} (and its variants) objects.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The fisher information matrix of \eqn{\Theta} in generalized linear models.
}
\description{
Compute or extract component-wise standard errors of Theta in multivariate (generalized) linear models of the form \deqn{Y = X\Theta + E} Useful to create confidence intervals and (multivariate) confidence regions under a Gaussian approximation of \eqn{\Theta}. Note that the Fisher information matrix is the one-data version (not scaled by the number of observations).
}
