% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNARfit-S3methods.R
\name{fitted.PLNARfit}
\alias{fitted.PLNARfit}
\title{Extracts model fitted values from objects returned by \code{\link[=PLN]{PLN()}} and its variants}
\usage{
\method{fitted}{PLNARfit}(object, ...)
}
\arguments{
\item{object}{an R6 object with class \code{\link{PLNARfit}}}

\item{...}{additional parameters for S3 compatibility. Not used}
}
\value{
A matrix of Fitted values extracted from the object object.
}
\description{
Extracts model fitted values from objects returned by \code{\link[=PLN]{PLN()}} and its variants
}
