% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNfit-S3methods.R
\name{fitted.PLNfit}
\alias{fitted.PLNfit}
\title{Extracts model fitted values from objects returned by \code{\link[=PLN]{PLN()}} and its variants}
\usage{
\method{fitted}{PLNfit}(object, ...)
}
\arguments{
\item{object}{an R6 object with class \code{\link{PLNfit}}}

\item{...}{additional parameters for S3 compatibility. Not used}
}
\value{
A matrix of Fitted values extracted from the object object.
}
\description{
Extracts model fitted values from objects returned by \code{\link[=PLN]{PLN()}} and its variants
}
