% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNmixturefit-S3methods.R
\name{fitted.PLNmixturefit}
\alias{fitted.PLNmixturefit}
\title{Extracts model fitted values from objects returned by \code{\link[=PLNmixture]{PLNmixture()}} and its variants}
\usage{
\method{fitted}{PLNmixturefit}(object, ...)
}
\arguments{
\item{object}{an R6 object with class \code{\link{PLNmixturefit}}}

\item{...}{additional parameters for S3 compatibility. Not used}
}
\value{
A matrix of Fitted values extracted from the object object.
}
\description{
Extracts model fitted values from objects returned by \code{\link[=PLNmixture]{PLNmixture()}} and its variants
}
