% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNfamily-class.R
\name{getBestModel.PLNfamily}
\alias{getBestModel.PLNfamily}
\title{Best model extraction from a collection of PLNfit}
\usage{
getBestModel.PLNfamily(object, crit = c("BIC", "ICL", "EBIC", "loglik",
  "R_squared"))
}
\arguments{
\item{object}{an R6 object with class PLNfamily}

\item{crit}{a character for the criterion used to performed the selection. Either
"BIC", "EBIC", "ICL", "loglik", "R_squared". Default is "BIC".}
}
\value{
Send back a object with class \code{\link[=PLNfit]{PLNfit}}.
}
\description{
Best model extraction from a collection of PLNfit
}
