% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods.R
\name{getBestModel.PLNnetworkfamily}
\alias{getBestModel.PLNnetworkfamily}
\title{Best model extraction from a collection of PLNnetworkfit}
\usage{
\method{getBestModel}{PLNnetworkfamily}(Robject, crit = c("BIC",
  "loglik", "R_squared", "EBIC", "StARS"), ...)
}
\arguments{
\item{Robject}{an object with class PLNnetworkfamilly}

\item{crit}{a character for the criterion used to performed the selection. Either
"BIC", "EBIC", "StARS", "R_squared". Default is \code{BIC}. If StARS
(Stability Approach to Regularization Selection) is chosen and stability selection
 was not yet performed, the function will call the method stability_selection with default argument.}

\item{...}{additional parameters for StARS criterion. stability, a scalar indicating the target stability (= 1 - 2 beta) at which the network is selected. Default is \code{0.9}.}
}
\value{
Send back a object with class \code{\link[=PLNnetworkfit]{PLNnetworkfit}}.
}
\description{
Best model extraction from a collection of PLNnetworkfit
}
