% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNPCAfamily-class.R
\name{getBestModel.PLNPCAfamily}
\alias{getBestModel.PLNPCAfamily}
\title{Best model extraction from a collection of PLNPCAfit}
\usage{
getBestModel.PLNPCAfamily(x, crit = c("ICL", "BIC", "R_squared"),
  stability = 0.9)
}
\arguments{
\item{x}{an object with classi PLNPCAfamilly}

\item{crit}{a character for the criterion used to performed the selection. Either
"BIC", "ICL", "R_squared". Default is \code{ICL}.}
}
\value{
Send back a object with class \code{\link[=PLNPCAfit]{PLNPCAfit}}.
}
\description{
Best model extraction from a collection of PLNPCAfit
}
