% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNnetworkfamily-class.R
\name{getModel.PLNnetworkfamily}
\alias{getModel.PLNnetworkfamily}
\title{Model extraction from a collection of PLNnetwork models}
\usage{
\method{getModel}{PLNnetworkfamily}(Robject, var, index = NULL)
}
\arguments{
\item{Robject}{an R6 object with class PLNfamily}

\item{var}{value of the parameter (penalty for sparse network) that identifies the model to be extracted from the collection. If no exact match is found, the model with closest parameter value is returned with a warning.}

\item{index}{Integer index of the model to be returned. Only the first value is taken into account.}
}
\value{
Sends back a object with class \code{\link[=PLNnetworkfit]{PLNnetworkfit}}.
}
\description{
Model extraction from a collection of PLNnetwork models
}
