% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNPCAfamily-class.R
\name{getModel.PLNPCAfamily}
\alias{getModel.PLNPCAfamily}
\title{Model extraction from a collection of PLNPCA models}
\usage{
\method{getModel}{PLNPCAfamily}(Robject, var, index = NULL)
}
\arguments{
\item{Robject}{an R6 object with class PLNfamily}

\item{var}{value of the parameter (rank for PCA) that identifies the model to be extracted from the collection. If no exact match is found, the model with closest parameter value is returned with a warning.}

\item{index}{Integer index of the model to be returned. Only the first value is taken into account.}
}
\value{
Sends back a object with class \code{\link[=PLNPCAfit]{PLNPCAfit}}.
}
\description{
Model extraction from a collection of PLNPCA models
}
