% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNfit-S3methods.R, R/ZIPLNfit-S3methods.R
\name{ICL}
\alias{ICL}
\alias{ICL.PLNfit}
\alias{ICL.ZIPLNfit}
\title{Integrated Classification Likelihood}
\usage{
ICL(object, ...)

\method{ICL}{PLNfit}(object, ...)

\method{ICL}{ZIPLNfit}(object, ...)
}
\arguments{
\item{object}{an R6 object with class \code{\link{ZIPLNfit}}}

\item{...}{additional parameters passed to methods}
}
\value{
A scalar: the variational ICL (larger is better).
}
\description{
Generic function to compute the Integrated Classification Likelihood (ICL) of a fitted model.
ICL = BIC - entropy of the variational distribution (larger is better).

\code{ICL.PLNfit}: ICL for a fitted \code{\link{PLNfit}}.

\code{ICL.ZIPLNfit}: ICL for a fitted \code{\link{ZIPLNfit}}.
}
\examples{
data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
model <- PLN(Abundance ~ 1, data = trichoptera)
ICL(model)
data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
model <- ZIPLN(Abundance ~ 1, data = trichoptera)
ICL(model)
}
