% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNfit-S3methods.R
\name{logLik.PLNfit}
\alias{logLik.PLNfit}
\title{Extract log-likelihood of a fitted PLN model}
\usage{
\method{logLik}{PLNfit}(object, ...)
}
\arguments{
\item{object}{an R6 object with class \code{\link{PLNfit}}}

\item{...}{additional parameters for S3 compatibility. Not used}
}
\value{
An object of class \code{"logLik"}. The numeric value is the variational ELBO.
Attributes \code{df} and \code{nobs} hold the number of parameters and observations.
}
\description{
Returns the variational lower bound of the log-likelihood as a \code{"logLik"} object,
compatible with \code{\link[stats:AIC]{stats::AIC()}} and \code{\link[stats:BIC]{stats::BIC()}}.
}
\examples{
data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
model <- PLN(Abundance ~ 1, data = trichoptera)
logLik(model)
}
