% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim-zipln.R
\name{optim_zipln_B}
\alias{optim_zipln_B}
\title{#' @importFrom glassoFast glassoFast
optim_zipln_Omega_sparse <- function(M, X, B, S, rho) {
n <- nrow(M); p <- ncol(M)
glassoFast::glassoFast( crossprod(M - X \%*\% B)/n + diag(colMeans(S * S), p, p), rho = rho )$wi
}}
\usage{
optim_zipln_B(M, X, Omega, config)
}
\description{
#' @importFrom glassoFast glassoFast
optim_zipln_Omega_sparse <- function(M, X, B, S, rho) {
n <- nrow(M); p <- ncol(M)
glassoFast::glassoFast( crossprod(M - X \%*\% B)/n + diag(colMeans(S * S), p, p), rho = rho )$wi
}
}
