% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLN.R
\name{PLN}
\alias{PLN}
\alias{PLN.formula}
\alias{PLN.default}
\title{Fit a Poisson lognormal model}
\usage{
PLN(Robject, ...)

\method{PLN}{formula}(formula, control = list())

\method{PLN}{default}(Y, X = cbind(rep(1, nrow(Y))), O = matrix(0, nrow(Y),
  ncol(Y)), control = list())
}
\arguments{
\item{Robject}{an R object, either a formula or a matrix}

\item{...}{additional parameters. Not used}

\item{formula}{a formula}

\item{control}{a list for controling the optimization. See details.}

\item{Y}{a (n x p) matrix of count data}

\item{X}{an optional (n x d) matrix of covariates. SHould include the intercept (a column of one) if the default method is used.}

\item{O}{an optional (n x p) matrix of offsets.}
}
\value{
an R6 object with class \code{\link[=PLNfit-class]{PLNfit}}
}
\description{
two methods are available for specifing the models (with formulas or matrices)
}
\details{
The parameter \code{control} is a list with the following entries
\itemize{
 \item{"factr"}{controls the L-BFGF-B procedure. See the documentation of \code{\link{optim}}. Default 1e8. Decrease if you experience instability or non monotonous J as a function of the rank}
 \item{"pgtol"}{controls the L-BFGF-B procedure. See the documentation of \code{\link{optim}}. Default 1e-2. Decrease if you experience instability or non monotonous J as a function of the rank}
 \item{"maxit"}{controls the L-BFGF-B procedure. See the documentaiton of \code{\link{optim}}. Default is 20000}
 \item{"lb.var"}{the minimum admissible value fr the variance parameter S in the variational approximation. Default is 1e-3.}
 \item{"cores"}{the number of cores. If Q has many entries, you might consider multiple cores. Default is 1.}
 \item{"trace"}{integer for verbosity. Useless when \code{cores} > 1}
}
}
\examples{
## See the vignette TODO!!!
}
\seealso{
The class  \code{\link[=PLNfit-class]{PLNfit}}
}
