% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNblock.R
\name{PLNblock}
\alias{PLNblock}
\title{Poisson lognormal model with blockwise covariance}
\usage{
PLNblock(
  formula,
  nb_blocks = 1:5,
  sparsity = 0,
  data,
  subset,
  weights,
  control = PLNblock_param()
)
}
\arguments{
\item{formula}{an object of class "formula": a symbolic description of the model to be fitted.}

\item{nb_blocks}{vector of number of blocks for regrouping the variables/responses (dimension of the residual covariance)}

\item{sparsity}{tuning parameter for controlling the sparsity level of Omega/Sigma either a scalar or a vector with the same size as \code{nb_blocks}}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which PLN is called.}

\item{subset}{an optional vector specifying a subset of observations to be used in the fitting process.}

\item{weights}{an optional vector of observation weights to be used in the fitting process.}

\item{control}{a list-like structure for controlling the optimization, with default generated by \code{\link[=PLN_param]{PLN_param()}}. See the associated documentation
for details.}
}
\value{
an R6 object with class \code{\link{PLNblockfamily}}, which contains
a collection of models with class \code{\link{PLNblockfit}}
}
\description{
Fit the multivariate Poisson lognormal with model with a variational algorithm. Use the (g)lm syntax for model specification (covariates, offsets, weights).
}
\examples{
data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPLN <- PLNblock(Abundance ~ 1, nb_blocks = 1:17, data = trichoptera)
}
\seealso{
The classes \code{\link{PLNblockfamily}} and \code{\link{PLNblockfit}}, and the and the configuration function \code{\link[=PLNblock_param]{PLNblock_param()}}.
}
