% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNblockbisfit-class.R
\name{PLNblockbisfit_sparse}
\alias{PLNblockbisfit_sparse}
\title{An R6 Class to represent a PLNfit with sparse block-wise residual covariance}
\description{
An R6 Class to represent a PLNfit with sparse block-wise residual covariance

An R6 Class to represent a PLNfit with sparse block-wise residual covariance
}
\examples{
\dontrun{
data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPLN <- PLNblockbis(Abundance ~ 1, data = trichoptera, nb_blocks = 1:5)
class(myPLN)
print(myPLN)
}
}
\section{Super classes}{
\code{\link[PLNmodels:PLNfit]{PLNmodels::PLNfit}} -> \code{\link[PLNmodels:PLNblockbisfit]{PLNmodels::PLNblockbisfit}} -> \code{PLNblockbisfit_sparse}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{penalty}}{tuning parameter for controlling the sparsity level of Omega/Sigma}

\item{\code{penalty_weights}}{a matrix of weights controlling the amount of penalty element-wise.}

\item{\code{n_edges}}{number of edges if the network (non null coefficient of the sparse precision matrix)}

\item{\code{nb_param}}{number of parameters in the current PLN model}

\item{\code{density}}{proportion of non-null edges in the network}

\item{\code{pen_loglik}}{variational lower bound of the l1-penalized loglikelihood}

\item{\code{EBIC}}{variational lower bound of the EBIC}

\item{\code{criteria}}{a vector with loglik, penalized loglik, BIC, EBIC, ICL, number of parameters, number of edges and graph density}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PLNblockbisfit_sparse-new}{\code{PLNblockbisfit_sparse$new()}}
\item \href{#method-PLNblockbisfit_sparse-optimize}{\code{PLNblockbisfit_sparse$optimize()}}
\item \href{#method-PLNblockbisfit_sparse-latent_network}{\code{PLNblockbisfit_sparse$latent_network()}}
\item \href{#method-PLNblockbisfit_sparse-plot_network}{\code{PLNblockbisfit_sparse$plot_network()}}
\item \href{#method-PLNblockbisfit_sparse-clone}{\code{PLNblockbisfit_sparse$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="optimize_vestep"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-optimize_vestep'><code>PLNmodels::PLNfit$optimize_vestep()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="postTreatment"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-postTreatment'><code>PLNmodels::PLNfit$postTreatment()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="predict"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-predict'><code>PLNmodels::PLNfit$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="predict_cond"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-predict_cond'><code>PLNmodels::PLNfit$predict_cond()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="print"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-print'><code>PLNmodels::PLNfit$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="show"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-show'><code>PLNmodels::PLNfit$show()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNblockbisfit" data-id="update"><a href='../../PLNmodels/html/PLNblockbisfit.html#method-PLNblockbisfit-update'><code>PLNmodels::PLNblockbisfit$update()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PLNblockbisfit_sparse-new"></a>}}
\if{latex}{\out{\hypertarget{method-PLNblockbisfit_sparse-new}{}}}
\subsection{Method \code{new()}}{
Initialize a \code{\link{PLNblockbisfit_sparse}} model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PLNblockbisfit_sparse$new(
  blocks,
  sparsity,
  responses,
  covariates,
  offsets,
  weights,
  formula,
  control
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{blocks}}{blocks indicators for regrouping the variables/responses (dimension of the residual covariance)}

\item{\code{sparsity}}{tuning parameter for controlling the sparsity level of Omega/Sigma}

\item{\code{responses}}{the matrix of responses (called Y in the model). Will usually be extracted from the corresponding field in PLNfamily-class}

\item{\code{covariates}}{design matrix (called X in the model). Will usually be extracted from the corresponding field in PLNfamily-class}

\item{\code{offsets}}{offset matrix (called O in the model). Will usually be extracted from the corresponding field in PLNfamily-class}

\item{\code{weights}}{an optional vector of observation weights to be used in the fitting process.}

\item{\code{formula}}{model formula used for fitting, extracted from the formula in the upper-level call}

\item{\code{control}}{a list for controlling the optimization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PLNblockbisfit_sparse-optimize"></a>}}
\if{latex}{\out{\hypertarget{method-PLNblockbisfit_sparse-optimize}{}}}
\subsection{Method \code{optimize()}}{
Call to the NLopt or TORCH optimizer and update of the relevant fields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PLNblockbisfit_sparse$optimize(responses, covariates, offsets, weights, config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{responses}}{the matrix of responses (called Y in the model). Will usually be extracted from the corresponding field in PLNfamily-class}

\item{\code{covariates}}{design matrix (called X in the model). Will usually be extracted from the corresponding field in PLNfamily-class}

\item{\code{offsets}}{offset matrix (called O in the model). Will usually be extracted from the corresponding field in PLNfamily-class}

\item{\code{weights}}{an optional vector of observation weights to be used in the fitting process.}

\item{\code{config}}{part of the \code{control} argument which configures the optimizer}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PLNblockbisfit_sparse-latent_network"></a>}}
\if{latex}{\out{\hypertarget{method-PLNblockbisfit_sparse-latent_network}{}}}
\subsection{Method \code{latent_network()}}{
Extract interaction network in the latent space
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PLNblockbisfit_sparse$latent_network(
  type = c("partial_cor", "support", "precision")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{edge value in the network. Can be "support" (binary edges), "precision" (coefficient of the precision matrix) or "partial_cor" (partial correlation between species)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a square matrix of size \code{PLNnetworkfit$n}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PLNblockbisfit_sparse-plot_network"></a>}}
\if{latex}{\out{\hypertarget{method-PLNblockbisfit_sparse-plot_network}{}}}
\subsection{Method \code{plot_network()}}{
plot the latent network.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PLNblockbisfit_sparse$plot_network(
  type = c("partial_cor", "support"),
  output = c("igraph", "corrplot"),
  edge.color = c("#F8766D", "#00BFC4"),
  remove.isolated = FALSE,
  node.labels = NULL,
  layout = layout_in_circle,
  plot = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{edge value in the network. Either "precision" (coefficient of the precision matrix) or "partial_cor" (partial correlation between species).}

\item{\code{output}}{Output type. Either \code{igraph} (for the network) or \code{corrplot} (for the adjacency matrix)}

\item{\code{edge.color}}{Length 2 color vector. Color for positive/negative edges. Default is \code{c("#F8766D", "#00BFC4")}. Only relevant for igraph output.}

\item{\code{remove.isolated}}{if \code{TRUE}, isolated node are remove before plotting. Only relevant for igraph output.}

\item{\code{node.labels}}{vector of character. The labels of the nodes. The default will use the column names ot the response matrix.}

\item{\code{layout}}{an optional igraph layout. Only relevant for igraph output.}

\item{\code{plot}}{logical. Should the final network be displayed or only sent back to the user. Default is \code{TRUE}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PLNblockbisfit_sparse-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PLNblockbisfit_sparse-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PLNblockbisfit_sparse$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
