% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNblockfamily-class.R
\name{PLNblockfamily}
\alias{PLNblockfamily}
\title{An R6 Class to represent a collection of PLNblockfit}
\description{
The function \code{\link[=PLNblock]{PLNblock()}} produces an instance of this class.

This class comes with a set of methods, some of them being useful for the user:
See the documentation for \code{\link[=getBestModel]{getBestModel()}},
\code{\link[=getModel]{getModel()}} and \link[=plot.PLNnetworkfamily]{plot()}
}
\examples{
data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
fits <- PLNblock(Abundance ~ 1, nb_blocks = 1:17, data = trichoptera)
class(fits)
}
\seealso{
The function \code{\link[=PLNblock]{PLNblock()}}, the class \code{\link{PLNblockfit}}
}
\section{Super class}{
\code{\link[PLNmodels:PLNfamily]{PLNmodels::PLNfamily}} -> \code{PLNblockfamily}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{nb_blocks}}{vector of number of blocks for regrouping the variables/responses (dimension of the residual covariance)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PLNblockfamily-new}{\code{PLNblockfamily$new()}}
\item \href{#method-PLNblockfamily-optimize}{\code{PLNblockfamily$optimize()}}
\item \href{#method-PLNblockfamily-getBestModel}{\code{PLNblockfamily$getBestModel()}}
\item \href{#method-PLNblockfamily-plot}{\code{PLNblockfamily$plot()}}
\item \href{#method-PLNblockfamily-plot_objective}{\code{PLNblockfamily$plot_objective()}}
\item \href{#method-PLNblockfamily-show}{\code{PLNblockfamily$show()}}
\item \href{#method-PLNblockfamily-clone}{\code{PLNblockfamily$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfamily" data-id="getModel"><a href='../../PLNmodels/html/PLNfamily.html#method-PLNfamily-getModel'><code>PLNmodels::PLNfamily$getModel()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfamily" data-id="postTreatment"><a href='../../PLNmodels/html/PLNfamily.html#method-PLNfamily-postTreatment'><code>PLNmodels::PLNfamily$postTreatment()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfamily" data-id="print"><a href='../../PLNmodels/html/PLNfamily.html#method-PLNfamily-print'><code>PLNmodels::PLNfamily$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PLNblockfamily-new"></a>}}
\if{latex}{\out{\hypertarget{method-PLNblockfamily-new}{}}}
\subsection{Method \code{new()}}{
Initialize all models in the collection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PLNblockfamily$new(
  nb_blocks,
  sparsity,
  responses,
  covariates,
  offsets,
  weights,
  formula,
  control
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nb_blocks}}{a vector of positive real number controlling the level of sparsity of the underlying network.}

\item{\code{sparsity}}{tuning parameter for controlling the sparsity level of Omega/Sigma}

\item{\code{responses}}{the matrix of responses common to every models}

\item{\code{covariates}}{the matrix of covariates common to every models}

\item{\code{offsets}}{the matrix of offsets common to every models}

\item{\code{weights}}{the vector of observation weights}

\item{\code{formula}}{model formula used for fitting, extracted from the formula in the upper-level call}

\item{\code{control}}{a list for controlling the optimization.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Update current \code{\link{PLNblockfit}} with smart starting values
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PLNblockfamily-optimize"></a>}}
\if{latex}{\out{\hypertarget{method-PLNblockfamily-optimize}{}}}
\subsection{Method \code{optimize()}}{
Call to the C++ optimizer on all models of the collection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PLNblockfamily$optimize(config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{a list for controlling the optimization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PLNblockfamily-getBestModel"></a>}}
\if{latex}{\out{\hypertarget{method-PLNblockfamily-getBestModel}{}}}
\subsection{Method \code{getBestModel()}}{
Extract best model in the collection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PLNblockfamily$getBestModel(crit = c("BIC", "ICL", "loglik"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{crit}}{a character for the criterion used to performed the selection. Either
"BIC", "ICL" or "loglik". Default is \code{ICL}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a \code{\link{PLNblockfit}} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PLNblockfamily-plot"></a>}}
\if{latex}{\out{\hypertarget{method-PLNblockfamily-plot}{}}}
\subsection{Method \code{plot()}}{
Display various outputs (goodness-of-fit criteria, robustness, diagnostic) associated with a collection of PLNblocks fits (a \code{\link{PLNblockfamily}})
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PLNblockfamily$plot(criteria = c("loglik", "ICL", "BIC"), reverse = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{criteria}}{vector of characters. The criteria to plot in \code{c("loglik", "BIC", "ICL")}. Defaults to all of them.}

\item{\code{reverse}}{A logical indicating whether to plot the value of the criteria in the "natural" direction
(loglik - 0.5 penalty) or in the "reverse" direction (-2 loglik + penalty). Default to FALSE, i.e use the
natural direction, on the same scale as the log-likelihood..}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a \code{\link{ggplot}} graph
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PLNblockfamily-plot_objective"></a>}}
\if{latex}{\out{\hypertarget{method-PLNblockfamily-plot_objective}{}}}
\subsection{Method \code{plot_objective()}}{
Plot objective value of the optimization problem along the penalty path
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PLNblockfamily$plot_objective()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a \code{\link{ggplot}} graph
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PLNblockfamily-show"></a>}}
\if{latex}{\out{\hypertarget{method-PLNblockfamily-show}{}}}
\subsection{Method \code{show()}}{
User friendly print method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PLNblockfamily$show()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PLNblockfamily-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PLNblockfamily-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PLNblockfamily$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
