% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNfamily-class.R
\name{PLNfamily_getBestModel}
\alias{PLNfamily_getBestModel}
\title{Best model extraction from a collection of PLNfit}
\arguments{
\item{crit}{a character for the criterion used to performed the selection. Either
"ICL", "BIC", "J" or "R2". Default is "ICL.}
}
\value{
Send back a object with class \code{\link[=PLNfit.PCA-class]{PLNfit}}.
}
\description{
Best model extraction from a collection of PLNfit
}

