% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNfamily-class.R
\name{PLNfamily_plot}
\alias{PLNfamily_plot}
\title{A plot method for a collection of PLNfit}
\value{
Produces a plot  representing the evolution of the criteria of the different models considered,
highlighting the best model in terms of ICL. If not smooth, you may consider running PLNPCA again with
a smaller tolerance for convergence.
}
\description{
A plot method for a collection of PLNfit
}

