% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNfamily-class.R
\name{PLNfamily_predict}
\alias{PLNfamily_predict}
\title{Predict counts of new samples for all fits in the family}
\arguments{
\item{newdata}{A optional data frame in which to look for variables with which to predict. If omitted, the family-level covariates are used.}

\item{newOffsets}{A optional matrix in which to look for offsets with which to predict. If omitted, the family-level offsets are used.}

\item{type}{The type of prediction required. The default is on the scale of the linear predictors (i.e. log average count);
the alternative "response" is on the scale of the response variable (i.e. average count)}
}
\value{
A list of matrices (one per fit) of predicted log-counts (if type = "link")
        or predicted counts (if type = "response").
}
\description{
Predict counts of new samples for all fits in the family
}
