% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNfamily-class.R
\name{PLNfamily_setCriteria}
\alias{PLNfamily_setCriteria}
\title{Set data frame with criteria (BIC, ICL, J and possibly R2) associated with the collection of fits}
\description{
Set data frame with criteria (BIC, ICL, J and possibly R2) associated with the collection of fits
}
