% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNfamily-class.R
\name{PLNfamily_show}
\alias{PLNfamily_show}
\title{Basic show method}
\arguments{
\item{verbose}{a logical (defaults to TRUE) controlling the amount of screen output.}
}
\description{
Basic show method
}
