% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNfit-class.R
\name{PLNfit_addCriteria}
\alias{PLNfit_addCriteria}
\title{Add (or update if existing) criteria value to a fit. New criteria are added at the end.}
\arguments{
\item{crit.name}{criteria name (given as a character string)}

\item{value}{criteria value (numeric)}
}
\description{
Add (or update if existing) criteria value to a fit. New criteria are added at the end.
}
