% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNfit-class.R
\name{PLNfit_diagonal}
\alias{PLNfit_diagonal}
\title{An R6 Class to represent a PLNfit in a standard, general framework, with diagonal residual covariance}
\description{
An R6 Class to represent a PLNfit in a standard, general framework, with diagonal residual covariance
}
\examples{
\dontrun{
data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPLN <- PLN(Abundance ~ 1, data = trichoptera)
class(myPLN)
print(myPLN)
}
}
\section{Super class}{
\code{\link[PLNmodels:PLNfit]{PLNfit}} -> \code{PLNfit_diagonal}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{nb_param}}{number of parameters in the current PLN model}

    \item{\code{vcov_model}}{character: the model used for the residual covariance}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-PLNfit_diagonal-initialize}{\code{PLNfit_diagonal$new()}}
    \item \href{#method-PLNfit_diagonal-clone}{\code{PLNfit_diagonal$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="optimize"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-optimize'><code>PLNfit$optimize()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="optimize_vestep"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-optimize_vestep'><code>PLNfit$optimize_vestep()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="postTreatment"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-postTreatment'><code>PLNfit$postTreatment()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="predict"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-predict'><code>PLNfit$predict()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="predict_cond"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-predict_cond'><code>PLNfit$predict_cond()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="print"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-print'><code>PLNfit$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="show"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-show'><code>PLNfit$show()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="update"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-update'><code>PLNfit$update()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PLNfit_diagonal-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-PLNfit_diagonal-initialize}{}}}
\subsection{\code{PLNfit_diagonal$new()}}{
  Initialize a \code{\link{PLNfit}} model
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PLNfit_diagonal$new(responses, covariates, offsets, weights, formula, control)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{responses}}{the matrix of responses (called Y in the model). Will usually be extracted from the corresponding field in PLNfamily-class}
      \item{\code{covariates}}{design matrix (called X in the model). Will usually be extracted from the corresponding field in PLNfamily-class}
      \item{\code{offsets}}{offset matrix (called O in the model). Will usually be extracted from the corresponding field in PLNfamily-class}
      \item{\code{weights}}{an optional vector of observation weights to be used in the fitting process.}
      \item{\code{formula}}{model formula used for fitting, extracted from the formula in the upper-level call}
      \item{\code{control}}{a list for controlling the optimization. See details.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PLNfit_diagonal-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PLNfit_diagonal-clone}{}}}
\subsection{\code{PLNfit_diagonal$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PLNfit_diagonal$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
