% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNfit-class.R
\name{PLNfit_fixedcov}
\alias{PLNfit_fixedcov}
\title{An R6 Class to represent a PLNfit in a standard, general framework, with fixed (inverse) residual covariance}
\description{
The function \code{\link[=PLN]{PLN()}} fit a model which is an instance of a object with class \code{\link{PLNfit}}.
Objects produced by the functions \code{\link[=PLNnetwork]{PLNnetwork()}}, \code{\link[=PLNPCA]{PLNPCA()}}, \code{\link[=PLNmixture]{PLNmixture()}} and \code{\link[=PLNLDA]{PLNLDA()}} also enjoy the methods of \code{\link[=PLNfit]{PLNfit()}} by inheritance.

This class comes with a set of R6 methods, some of them being useful for the user and exported as S3 methods.
See the documentation for \code{\link[=coef]{coef()}}, \code{\link[=sigma]{sigma()}}, \code{\link[=predict]{predict()}}, \code{\link[=vcov]{vcov()}} and \code{\link[=standard_error]{standard_error()}}.

Fields are accessed via active binding and cannot be changed by the user.
}
\examples{
\dontrun{
data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPLN <- PLN(Abundance ~ 1, data = trichoptera)
class(myPLN)
print(myPLN)
}
}
\section{Super class}{
\code{\link[PLNmodels:PLNfit]{PLNmodels::PLNfit}} -> \code{PLNfit_fixedcov}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{nb_param}}{number of parameters in the current PLN model}

\item{\code{vcov_model}}{character: the model used for the residual covariance}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PLNfit_fixedcov-new}{\code{PLNfit_fixedcov$new()}}
\item \href{#method-PLNfit_fixedcov-optimize}{\code{PLNfit_fixedcov$optimize()}}
\item \href{#method-PLNfit_fixedcov-update_Sigma}{\code{PLNfit_fixedcov$update_Sigma()}}
\item \href{#method-PLNfit_fixedcov-update_loglik}{\code{PLNfit_fixedcov$update_loglik()}}
\item \href{#method-PLNfit_fixedcov-get_objective}{\code{PLNfit_fixedcov$get_objective()}}
\item \href{#method-PLNfit_fixedcov-clone}{\code{PLNfit_fixedcov$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="VEstep"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-VEstep'><code>PLNmodels::PLNfit$VEstep()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="get_vcov_hat"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-get_vcov_hat'><code>PLNmodels::PLNfit$get_vcov_hat()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="postTreatment"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-postTreatment'><code>PLNmodels::PLNfit$postTreatment()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="predict"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-predict'><code>PLNmodels::PLNfit$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="predict_cond"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-predict_cond'><code>PLNmodels::PLNfit$predict_cond()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="print"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-print'><code>PLNmodels::PLNfit$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="set_R2"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-set_R2'><code>PLNmodels::PLNfit$set_R2()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="show"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-show'><code>PLNmodels::PLNfit$show()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="torch_Sigma"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-torch_Sigma'><code>PLNmodels::PLNfit$torch_Sigma()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="torch_elbo"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-torch_elbo'><code>PLNmodels::PLNfit$torch_elbo()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="torch_optimize"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-torch_optimize'><code>PLNmodels::PLNfit$torch_optimize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="torch_vloglik"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-torch_vloglik'><code>PLNmodels::PLNfit$torch_vloglik()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="update"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-update'><code>PLNmodels::PLNfit$update()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="vcov_sandwich"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-vcov_sandwich'><code>PLNmodels::PLNfit$vcov_sandwich()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="vcov_wald"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-vcov_wald'><code>PLNmodels::PLNfit$vcov_wald()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PLNfit_fixedcov-new"></a>}}
\if{latex}{\out{\hypertarget{method-PLNfit_fixedcov-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PLNfit_fixedcov$new(responses, covariates, offsets, weights, formula, control)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PLNfit_fixedcov-optimize"></a>}}
\if{latex}{\out{\hypertarget{method-PLNfit_fixedcov-optimize}{}}}
\subsection{Method \code{optimize()}}{
Call to the NLopt or TORCH optimizer and update of the relevant fields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PLNfit_fixedcov$optimize(responses, covariates, offsets, weights, control)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PLNfit_fixedcov-update_Sigma"></a>}}
\if{latex}{\out{\hypertarget{method-PLNfit_fixedcov-update_Sigma}{}}}
\subsection{Method \code{update_Sigma()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PLNfit_fixedcov$update_Sigma(weights)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PLNfit_fixedcov-update_loglik"></a>}}
\if{latex}{\out{\hypertarget{method-PLNfit_fixedcov-update_loglik}{}}}
\subsection{Method \code{update_loglik()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PLNfit_fixedcov$update_loglik(weights, Y)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PLNfit_fixedcov-get_objective"></a>}}
\if{latex}{\out{\hypertarget{method-PLNfit_fixedcov-get_objective}{}}}
\subsection{Method \code{get_objective()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PLNfit_fixedcov$get_objective()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PLNfit_fixedcov-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PLNfit_fixedcov-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PLNfit_fixedcov$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
