% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNfit-class.R
\name{PLNfit_genpop}
\alias{PLNfit_genpop}
\title{An R6 Class to represent a PLNfit with a residual covariance structured by a fixed
correlation matrix (e.g. a genetic relationship matrix), motivated by population genetics}
\description{
Sigma = sigma2 * (rho * C + (1 - rho) * I_p), where C is a fixed p x p
correlation matrix supplied by the user (\code{control$C}) and (sigma2, rho) are estimated.
See \code{GeneticCovTraits} in \code{src/covariance_pln.h} for the C++ side.
}
\examples{
\dontrun{
data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
p <- ncol(trichoptera$Abundance)
C <- 0.5^abs(outer(1:p, 1:p, "-")); diag(C) <- 1
myPLN <- PLN(Abundance ~ 1, data = trichoptera, control = PLN_param(covariance = "genpop", C = C))
class(myPLN)
print(myPLN)
}
}
\section{Super class}{
\code{\link[PLNmodels:PLNfit]{PLNfit}} -> \code{PLNfit_genpop}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{nb_param}}{number of parameters in the current PLN model}

    \item{\code{vcov_model}}{character: the model used for the residual covariance}

    \item{\code{gen_par}}{a list with the two extra parameters of the genpop covariance model:
sigma2 (variance scale) and rho (mixing weight / heritability), decoded from Sigma and C.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-PLNfit_genpop-initialize}{\code{PLNfit_genpop$new()}}
    \item \href{#method-PLNfit_genpop-clone}{\code{PLNfit_genpop$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="optimize"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-optimize'><code>PLNfit$optimize()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="optimize_vestep"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-optimize_vestep'><code>PLNfit$optimize_vestep()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="postTreatment"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-postTreatment'><code>PLNfit$postTreatment()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="predict"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-predict'><code>PLNfit$predict()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="predict_cond"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-predict_cond'><code>PLNfit$predict_cond()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="print"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-print'><code>PLNfit$print()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="show"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-show'><code>PLNfit$show()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="update"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-update'><code>PLNfit$update()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PLNfit_genpop-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-PLNfit_genpop-initialize}{}}}
\subsection{\code{PLNfit_genpop$new()}}{
  Initialize a \code{\link{PLNfit_genpop}} model
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PLNfit_genpop$new(responses, covariates, offsets, weights, formula, control)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{responses}}{the matrix of responses (called Y in the model). Will usually be extracted from the corresponding field in PLNfamily-class}
      \item{\code{covariates}}{design matrix (called X in the model). Will usually be extracted from the corresponding field in PLNfamily-class}
      \item{\code{offsets}}{offset matrix (called O in the model). Will usually be extracted from the corresponding field in PLNfamily-class}
      \item{\code{weights}}{an optional vector of observation weights to be used in the fitting process.}
      \item{\code{formula}}{model formula used for fitting, extracted from the formula in the upper-level call}
      \item{\code{control}}{a list for controlling the optimization, must include a field \code{C} (the fixed p x p correlation matrix). See details.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PLNfit_genpop-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PLNfit_genpop-clone}{}}}
\subsection{\code{PLNfit_genpop$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{PLNfit_genpop$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
