% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNfit-class.R, R/PLNPCAfit-class.R
\name{PLNfit_latentPos}
\alias{PLNfit_latentPos}
\alias{PLNfit_latentPos}
\title{Positions in the (euclidian) parameter space, noted as Z in the model. Used to compute the likelihood.}
\arguments{
\item{covariates}{a matrix of covariates. Will usually be extracted from the corresponding field in PLNfamily-class}

\item{offsets}{a matrix of offsets. Will usually be extracted from the corresponding field in PLNfamily-class}

\item{covariates}{a matrix of covariates. Will usually be extracted from the corresponding field in PLNfamily-class}

\item{offsets}{a matrix of offsets. Will usually be extracted from the corresponding field in PLNfamily-class}
}
\description{
Positions in the (euclidian) parameter space, noted as Z in the model. Used to compute the likelihood.

Positions in the (euclidian) parameter space, noted as Z in the model. Used to compute the likelihood.
}
