% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNfit-class.R
\name{PLNfit_plot}
\alias{PLNfit_plot}
\title{Display the model parameters in a matrix fashion}
\arguments{
\item{type}{character. should the variational or the model parameters be plotted? default is "model".}
}
\description{
Display the model parameters in a matrix fashion
}
