% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNLDA.R
\name{PLNLDA}
\alias{PLNLDA}
\title{Poisson lognormal model towards Linear Discriminant Analysis}
\usage{
PLNLDA(formula, data, subset, weights, grouping, control = PLNLDA_param())
}
\arguments{
\item{formula}{an object of class "formula": a symbolic description of the model to be fitted.}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which lm is called.}

\item{subset}{an optional vector specifying a subset of observations to be used in the fitting process.}

\item{weights}{an optional vector of observation weights to be used in the fitting process.}

\item{grouping}{a factor specifying the class of each observation used for discriminant analysis.}

\item{control}{a list-like structure for controlling the optimization, with default generated by \code{\link[=PLNLDA_param]{PLNLDA_param()}}. See the associated documentation.}
}
\value{
an R6 object with class \code{\link[=PLNLDAfit]{PLNLDAfit()}}
}
\description{
Fit the Poisson lognormal for LDA with a variational algorithm. Use the (g)lm syntax for model specification (covariates, offsets).
}
\details{
See \code{\link[=PLNLDA_param]{PLNLDA_param()}} for a full description of the optimization parameters.
Note that unlike \code{\link[=PLN_param]{PLN_param()}}, \code{\link[=PLNLDA_param]{PLNLDA_param()}} does not expose the \code{"fixed"} covariance option or the \code{Omega} parameter, which are not meaningful in the LDA context.
}
\examples{
data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPLNLDA <- PLNLDA(Abundance ~ 1, grouping = Group, data = trichoptera)
}
\seealso{
The class \code{\link{PLNLDAfit}}
}
