% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNLDAfit-class.R
\name{PLNLDAfit_plot_LDA}
\alias{PLNLDAfit_plot_LDA}
\title{Plot a summary of the current \code{PLNLDAfit} object}
\arguments{
\item{axes}{numeric a vector of axes to be considered. The default is 1:min(3,rank).}

\item{plot}{logical. Should the plot be displayed or sent back (ggplot object)}

\item{var.cols}{a character, factor or numeric to define the color associated with the variables. Default is "gray"}
}
\value{
a plot with a matrix-like layout with size nb.axes x nb.axes, displaying individual maps and correlation circles for the corresponding axes
}
\description{
Plot a summary of the current \code{PLNLDAfit} object
}
