% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNmodels-package.R
\docType{package}
\name{PLNmodels-package}
\alias{PLNmodels}
\alias{PLNmodels-package}
\title{PLNmodels: Poisson Lognormal Models}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The Poisson-lognormal model and variants (Chiquet, Mariadassou and Robin, 2021 \doi{10.3389/fevo.2021.588292}) can be used for a variety of multivariate problems when count data are at play, including principal component analysis for count data, discriminant analysis, model-based clustering and network inference. Implements variational algorithms to fit such models accompanied with a set of functions for visualization and diagnostic.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pln-team.github.io/PLNmodels/}
  \item Report bugs at \url{https://github.com/pln-team/PLNmodels/issues}
}

}
\author{
\strong{Maintainer}: Julien Chiquet \email{julien.chiquet@inrae.fr} (\href{https://orcid.org/0000-0002-3629-3429}{ORCID})

Authors:
\itemize{
  \item Julien Chiquet \email{julien.chiquet@inrae.fr} (\href{https://orcid.org/0000-0002-3629-3429}{ORCID})
  \item Mahendra Mariadassou \email{mahendra.mariadassou@inrae.fr} (\href{https://orcid.org/0000-0003-2986-354X}{ORCID})
  \item Stéphane Robin \email{stephane.robin@inrae.fr}
}

Other contributors:
\itemize{
  \item François Gindraud \email{francois.gindraud@gmail.com} [contributor]
  \item Julie Aubert \email{julie.aubert@inrae.fr} [contributor]
  \item Bastien Batardière \email{bastien.batardiere@inrae.fr} [contributor]
  \item Giovanni Poggiato \email{giov.poggiato@gmail.com} [contributor]
  \item Cole Trapnell \email{coletrap@uw.edu} [contributor]
  \item Maddy Duran \email{duran@uw.edu} [contributor]
  \item Ryan Friedman \email{ryanfriedman22@gmail.com} [contributor]
}

}
\keyword{internal}
