% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNmodels-package.R
\docType{package}
\name{PLNmodels}
\alias{PLNmodels}
\alias{PLNmodels-package}
\title{PLNmodels}
\description{
Implements probabilistic PCA for count data via a Poisson log-normal model. Can use covariates
and offset on top of the main observation matrix in a GLM setup. The algorithm relies on variational
inference which is performed by maximizing a bi-convex function with the L-BFGS-B method.
}
\details{
Try vignette("trichoptera", package="PLNmodels") for an (not yet comprehensive) example.

The main function is \code{\link{PLNPCA}} which produces two kinds of Reference Class objects,
\code{\link[=PLNfamily-class]{PLNfamily}} and \code{\link[=PLNfit.PCA-class]{PLNfit.PCA}}. See their
documentation and the associated methods for manipulation.
}
\author{
Julien Chiquet \email{julien.chiquet@inra.fr}

Mahendra Mariadassou \email{mahendra.mariadassou@inra.fr}

Stéphane Robin \email{stephane.robin@inra.fr}
}
