% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNnetwork.R
\name{PLNnetwork_stabs}
\alias{PLNnetwork_stabs}
\alias{PLNnetwork_stabs.formula}
\alias{PLNnetwork_stabs.default}
\title{Performs stability selection for Poisson lognormal models towards network inference}
\usage{
PLNnetwork_stabs(Robject, ...)

\method{PLNnetwork_stabs}{formula}(formula, penalties, approx = FALSE,
  control.init = list(), control.main = list())

\method{PLNnetwork_stabs}{default}(Y, X = cbind(rep(1, nrow(Y))),
  O = matrix(0, nrow(Y), ncol(Y)), penalties = NULL, approx = FALSE,
  control.init = list(), control.main = list())
}
\description{
two methods are available for specifying the models (with formulas or matrices)
}
