% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNnetworkfamily-class.R
\name{PLNnetworkfamily_optimize}
\alias{PLNnetworkfamily_optimize}
\title{One (coordinate-descent) optimization for each \code{\link[=PLNnetworkfit-class]{PLNnetworkfit}} model, using
the same starting point (inception model from \code{\link[=PLNfit-class]{PLNfit}}) for all models.}
\description{
One (coordinate-descent) optimization for each \code{\link[=PLNnetworkfit-class]{PLNnetworkfit}} model, using
the same starting point (inception model from \code{\link[=PLNfit-class]{PLNfit}}) for all models.
}
