% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNnetworkfamily-class.R
\name{PLNnetworkfamily_optimize_approx}
\alias{PLNnetworkfamily_optimize_approx}
\title{One fit for each \code{\link[=PLNnetworkfit-class]{PLNnetworkfit}} model, using
the same starting point (inception model from \code{\link[=PLNfit-class]{PLNfit}}) for all models.
Unlike \code{\link[=PLNnetworkfamily_optimize]{optimize}}, the optimization does not use an iterative procedure:
- Theta, M and S are fixed to their inception values
- Omega/Sigma is optimized only once using graphical lasso}
\description{
One fit for each \code{\link[=PLNnetworkfit-class]{PLNnetworkfit}} model, using
the same starting point (inception model from \code{\link[=PLNfit-class]{PLNfit}}) for all models.
Unlike \code{\link[=PLNnetworkfamily_optimize]{optimize}}, the optimization does not use an iterative procedure:
- Theta, M and S are fixed to their inception values
- Omega/Sigma is optimized only once using graphical lasso
}
