% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNnetworkfamily-class.R
\name{PLNnetworkfamily_plot}
\alias{PLNnetworkfamily_plot}
\title{A plot method for a collection of PLNnetworkfit}
\value{
Produces a plot  representing the evolution of the criteria of the different models considered,
highlighting the best model in terms of BIC.
}
\description{
A plot method for a collection of PLNnetworkfit
}
