% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNnetworkfamily-class.R
\name{PLNnetworkfamily_setPenalties}
\alias{PLNnetworkfamily_setPenalties}
\title{Set penalties in a \code{\link[=PLNnetworkfamily-class]{PLNnetworkfamily}} family
of \code{\link[=PLNnetworkfit-class]{PLNnetworkfit}} models}
\arguments{
\item{penalties}{A numeric vector with the sparsity levels of the networks in family. If NULL, a relevant
vector is automatically computed from the inception model.}

\item{nPenalties}{The number of penalties to use. A warning is thrown if this does not match the number of
models in the family.}

\item{verbose}{Logical. Controls the amount of screen output.}
}
\description{
Set penalties in a \code{\link[=PLNnetworkfamily-class]{PLNnetworkfamily}} family
of \code{\link[=PLNnetworkfit-class]{PLNnetworkfit}} models
}
