% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNPCAfamily-class.R
\docType{data}
\name{PLNPCAfamily}
\alias{PLNPCAfamily}
\title{A Reference Class to represent a collection of PLNPCAfit}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PLNPCAfamily
}
\description{
The function \code{\link{PLNPCA}} produces an instance of this class.

This class comes with a set of methods, some of them being useful for the user:
See the documentation for \code{\link[=PLNPCAfamily_getBestModel]{getBestModel}},
\code{\link[=PLNPCAfamily_getModel]{getModel}} and \code{\link[=PLNPCAfamily_plot]{plot}}. Other methods
 should not be called as they are designed to be used during the optimization process.

Fields should not be changed or manipulated by the user as they are updated internally
during the estimation process.
}
\section{Fields}{

\describe{
\item{\code{type}}{a character indicating the model used for the covariance matrix in the variational Gaussian approximation. Either "diagonal" or "spherical".}

\item{\code{ranks}}{the dimensions of the successively fitted models}

\item{\code{models}}{a list of \code{\link[=PLNPCAfit.PCA-class]{PLNPCAfit}} object, one per rank.}

\item{\code{criteria}}{a data frame with the value of some criteria (variational lower bound J, BIC, ICL and R2) for the different models.}

\item{\code{responses}}{the matrix of responses common to every models}

\item{\code{covariates}}{the matrix of covariates common to every models}

\item{\code{offsets}}{the matrix of offsets common to every models}

\item{\code{objective}}{the R function used to compute the model's objective during the optimization process}

\item{\code{gradient}}{the R function to compute the model's gradient during the optimization process}
}}

\seealso{
The function \code{\link{PLNPCA}}, the class \code{\link[=PLNPCAfit.PCA-class]{PLNPCAfit.PCA}}
}
\keyword{datasets}
