% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNPCAfamily-class.R
\name{PLNPCAfamily_getBestModel}
\alias{PLNPCAfamily_getBestModel}
\title{Best model extraction from a collection of PLNPCAfit}
\arguments{
\item{crit}{a character for the criterion used to performed the selection. Either
"ICL", "BIC", "J" or "R2". Default is "ICL.}
}
\value{
Send back a object with class \code{\link[=PLNPCAfit.PCA-class]{PLNPCAfit}}.
}
\description{
Best model extraction from a collection of PLNPCAfit
}
