% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNPCAfamily-class.R
\name{PLNPCAfamily_getModel}
\alias{PLNPCAfamily_getModel}
\title{Model extraction from a collection of PLNPCAfit}
\arguments{
\item{rank}{an integer given the rank of the model to be extracted from the collection.}
}
\value{
Send back a object with class \code{\link[=PLNPCAfit]{PLNPCAfit}}.
}
\description{
Model extraction from a collection of PLNPCAfit
}
