% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNPCAfit-class.R
\docType{data}
\name{PLNPCAfit}
\alias{PLNPCAfit}
\title{An R6 Class to represent a PLNfit in a PCA framework}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PLNPCAfit
}
\description{
The function \code{\link{PLNPCA}} produces a collection of models which are instances of object with class \code{PLNPCAfit}.

This class comes with a set of methods, some of them being useful for the user:
See the documentation for \code{\link[=PLNPCAfit_plot]{plot}}, \code{\link[=PLNPCAfit_plot_individual.map]{plot_individual.map}}
and \code{\link[=PLNPCAfit_plot_correlation.circle]{plot_correlation.circle}}

Fields should not be changed or manipulated by the user as they are updated internally.
}
\section{Fields}{

\describe{
\item{\code{type}}{a character indicating the model used for the covariance matrix in the variational Gaussian approximation. Either "diagonal" or "spherical".}

\item{\code{rank}}{the dimension of the current model}

\item{\code{model.par}}{a list with two matrices, B and Theta, which are the estimated parameters of the pPCA model}

\item{\code{variation.par}}{a list with two matrices, M and S, which are the estimated parameters in the variational approximation}

\item{\code{criteria}}{a named vector with the value of some criteria (variational lower bound J, BIC, ICL, R2, lmin and lmax) for the different models.}

\item{\code{convergence}}{numeric; the convergence status of the L-BFGS-B method}

\item{\code{loglik}}{numeric; the Poisson loglikelihood of the current model}

\item{\code{percentVar}}{the percent of variance explained by each axis}

\item{\code{corrCircle}}{a matrix of correlations to plot the correlation circles}

\item{\code{scores}}{a matrix of scores to plot the individual factor maps}
}}

\seealso{
The function \code{\link{PLNPCA}}, the class \code{\link[=PLNfamily-class]{PLNfamily}}
}
\keyword{datasets}
