% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNPCAfit-class.R
\name{PLNPCAfit_plot_correlation_circle}
\alias{PLNPCAfit_plot_correlation_circle}
\title{Plot the correlation circle of a specified axis for a \code{PLNPCAfit} object}
\arguments{
\item{axes}{numeric, the axes to use for the plot. Default it c(1,2)}

\item{main}{character, the title. Default is "Variable Factor map"}

\item{plot}{logical. Should the plot be displayed or sent back (ggplot object)}

\item{cols}{a character, factor or numeric to defined the color associated with the variable. Default is "gray"}

\item{size}{integer, the size of the labels}
}
\value{
displays a correlation circle for the corresponding axes and/or sends back a ggplot2 object
}
\description{
Plot the correlation circle of a specified axis for a \code{PLNPCAfit} object
}
