% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNPCAfit-class.R, R/PLNnetworkfit-class.R
\name{PLNPCAfit_plot_individual.map}
\alias{PLNPCAfit_plot_individual.map}
\alias{PLNPCAfit_plot_individual.map}
\title{Plot the individual map of a specified axis for a \code{PLNPCAfit} object}
\arguments{
\item{axes}{numeric, the axes to use for the plot. Default it c(1,2)}

\item{main}{character, the title. Default is "Individual Factor Map"}

\item{plot}{logical. Should the plot be displayed or sent back as a ggplot object}

\item{cols}{a character, factor or numeric to defined the color associated with the observations. Default is "gray"}

\item{plot}{logical. Should the plot be displayed or sent back as an igraph object}

\item{remove.isolated}{if \code{TRUE}, isolated node are remove before plotting.}

\item{layout}{an optional igraph layout}
}
\value{
displays a individual map for thecorresponding axes and/or sends back a ggplot2 object

displays a graph (via igraph for small graph and corrplot for large ones) and/or sends back an igraph object
}
\description{
Plot the individual map of a specified axis for a \code{PLNPCAfit} object

Plot the network (support of the inverse covariance) for a \code{PLNnetworkfit} object
}
