% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNfit-class.R
\name{plot.PLNfit}
\alias{plot.PLNfit}
\title{Display the model parameters in a matrix fashion}
\usage{
\method{plot}{PLNfit}(x, type = c("model", "variational"))
}
\arguments{
\item{x}{an R6 object with class PLNfit}

\item{type}{character. should the variational or the model parameters be plotted? default is "model".}
}
\description{
Display the model parameters in a matrix fashion
}
