% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNLDAfit-S3methods.R
\name{plot.PLNLDAfit}
\alias{plot.PLNLDAfit}
\title{LDA visualization (individual and/or variable factor map(s)) for a \code{\link{PLNPCAfit}} object}
\usage{
\method{plot}{PLNLDAfit}(
  x,
  map = c("both", "individual", "variable"),
  nb_axes = min(3, x$rank),
  axes = seq.int(min(2, x$rank)),
  var_cols = "var_colors",
  plot = TRUE,
  main = NULL,
  ...
)
}
\arguments{
\item{x}{an R6 object with class PLNPCAfit}

\item{map}{the type of output for the PCA visualization: either "individual", "variable" or "both". Default is "both".}

\item{nb_axes}{scalar: the number of axes to be considered when map = "both". The default is min(3,rank).}

\item{axes}{numeric, the axes to use for the plot when map = "individual" or "variable". Default it c(1,min(rank))}

\item{var_cols}{a character or factor to define the color associated with the variables. By default, all variables receive the default color of the current palette.}

\item{plot}{logical. Should the plot be displayed or sent back as \code{\link{ggplot2}} object}

\item{main}{character. A title for the single plot (individual or variable factor map). If NULL (the default), an hopefully appropriate title will be used.}

\item{...}{Not used (S3 compatibility).}
}
\value{
displays an individual and/or variable factor maps for the corresponding axes, and/or sends back a \code{\link{ggplot2}} or gtable object
}
\description{
LDA visualization (individual and/or variable factor map(s)) for a \code{\link{PLNPCAfit}} object
}
\examples{
data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPLNLDA <- PLNLDA(Abundance ~ 1, grouping = Group, data = trichoptera)
\dontrun{
plot(myPLNLDA, map = "individual", nb_axes = 2)
}
}
