% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNPCAfit-S3methods.R
\name{plot.PLNPCAfit}
\alias{plot.PLNPCAfit}
\title{Plot the individual or variable factor map for a \code{PLNPCAfit} object}
\usage{
\method{plot}{PLNPCAfit}(x, map = c("individual", "variable"),
  axes = 1:min(2, x$rank), plot = TRUE, ...)
}
\arguments{
\item{x}{an R6 object with class PLNPCAfit}

\item{axes}{numeric, the axes to use for the plot. Default it c(1,2)}

\item{plot}{logical. Should the plot be displayed or sent back (ggplot object)}

\item{...}{additional parameters: cols, main.}
}
\value{
displays a individual or variable map for the corresponding axes and/or sends back a ggplot2 object
}
\description{
Plot the individual or variable factor map for a \code{PLNPCAfit} object
}
