% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNnetworkfit-class.R
\name{plot_network}
\alias{plot_network}
\title{Plot the network (support of the inverse covariance) for a \code{PLNnetworkfit} object}
\arguments{
\item{plot}{logical. Should the plot be displayed or sent back as an igraph object}

\item{remove.isolated}{if \code{TRUE}, isolated node are remove before plotting.}

\item{layout}{an optional igraph layout}
}
\value{
displays a graph (via igraph for small graph and corrplot for large ones) and/or sends back an igraph object
}
\description{
Plot the network (support of the inverse covariance) for a \code{PLNnetworkfit} object
}
