% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNfit-class.R
\name{predict.PLNfit}
\alias{predict.PLNfit}
\title{Predict counts of a new sample}
\usage{
\method{predict}{PLNfit}(x, newdata, newOffsets, type = c("link", "response"))
}
\arguments{
\item{x}{an R6 object with class PLNfit}

\item{newdata}{A data frame in which to look for variables with which to predict.}

\item{newOffsets}{A matrix in which to look for offsets with which to predict.}

\item{type}{The type of prediction required. The default is on the scale of the linear predictors (i.e. log average count);
the alternative "response" is on the scale of the response variable (i.e. average count)}
}
\value{
A matrix of predicted log-counts (if type = "link") or predicted counts (if type = "response").
}
\description{
Predict counts of a new sample
}
