% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNLDAfit-class.R
\name{predict.PLNLDAfit}
\alias{predict.PLNLDAfit}
\title{Predict group of new samples}
\usage{
\method{predict}{PLNLDAfit}(object, newdata, newOffsets, newCounts,
  type = c("posterior", "response"), control = list(), ...)
}
\arguments{
\item{object}{an R6 object with class PLNLDAfit}

\item{newdata}{A data frame in which to look for variables with which to predict.}

\item{newOffsets}{A matrix in which to look for offsets with which to predict.}

\item{newCounts}{A matrix in which to look for counts with to predict}

\item{type}{The type of prediction required. The default are posterior probabilities for each group (in log-scale),
the alternative "response" is the group with maximal posterior probability.}

\item{control}{a list for controlling the optimization. See \code{\link[=PLN]{PLN}} for details.}

\item{...}{additional parameters for S3 compatibility. Not used}
}
\value{
A matrix of predicted scores for each group (if type = "score") or a vector of predicted
        groups (if type = "response").
}
\description{
Predict group of new samples
}
