% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIPLNfits-S3methods.R
\name{predict.ZIPLNfit}
\alias{predict.ZIPLNfit}
\title{Predict counts of a new sample}
\usage{
\method{predict}{ZIPLNfit}(object, newdata, type = c("link", "response"), ...)
}
\arguments{
\item{object}{an R6 object with class \code{\link{ZIPLNfit}}}

\item{newdata}{A data frame in which to look for variables and offsets with which to predict}

\item{type}{The type of prediction required. The default is on the scale of the linear predictors (i.e. log average count)}

\item{...}{additional parameters for S3 compatibility. Not used}
}
\value{
A matrix of predicted log-counts (if \code{type = "link"}) or predicted counts (if \code{type = "response"}).
}
\description{
Predict counts of a new sample
}
