% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNfit-S3methods.R
\name{predict_cond}
\alias{predict_cond}
\title{Predict counts conditionally}
\usage{
predict_cond(object, newdata, Yc, type = c("link", "response"))
}
\arguments{
\item{object}{an R6 object with class \code{\link{PLNfit}}}

\item{newdata}{A data frame in which to look for variables and offsets with which to predict}

\item{Yc}{a dataframe containing the abundances of the observed species (matching the names provided as data in the PLN function)}

\item{type}{The type of prediction required. The default is on the scale of the linear predictors (i.e. log average count)}

\item{...}{additional parameters for S3 compatibility. Not used}
}
\value{
A list containing:
\item{preds}{A matrix of predicted log-counts (if \code{type = "link"}) or predicted counts (if \code{type = "response"})}
\item{M}{A matrix containing E(Z_uncond | Y_c) for each given site.}
\item{S}{A matrix containing Var(Z_uncond | Y_c) for each given site (sites are the third dimension of the array)}
}
\description{
Predict counts of a new sample conditionally on a (set of) observed species
}
\examples{
data(trichoptera)
trichoptera_prep <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPLN <- PLN(Abundance ~ Temperature + Wind, trichoptera_prep)
#Condition on the set of the first two species in the dataset (Hym, Hys) at the ten first sites
Yc <- trichoptera$Abundance[1:10,c(1,2)]
newdata <- cbind(1,trichoptera$Covariate[1:10,c("Temperature", "Wind")])
pred <- predict_cond(object=myPLN,newdata=newdata,Yc=Yc,type="response")
}
