% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_utils.R
\name{prepare_data_from_biom}
\alias{prepare_data_from_biom}
\title{Prepare data for use in PLN models from a biom object}
\usage{
prepare_data_from_biom(biom, offset = "TSS", ...)
}
\arguments{
\item{biom}{Required. Either a biom-class object from which the count table and covariates data.frame are extracted or a file name where to read the biom.}

\item{offset}{Optional. Normalisation scheme used to compute scaling factors used as offset during PLN inference.}

\item{...}{Addtional arguments passed on to \code{\link[=compute_offset]{compute_offset}}}
}
\description{
Wrapper around \code{\link[=prepare_data]{prepare_data}}, extracts the count table and the covariates data.frame from a "biom" class object
before passing them to \code{\link[=prepare_data]{prepare_data}}. See \code{\link[=prepare_data]{prepare_data}} for details.
}
\seealso{
\code{\link[=compute_offset]{compute_offset}} and \code{\link[=prepare_data]{prepare_data}}
}
