% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIPLNfit-S3methods.R
\name{sigma.ZIPLNfit}
\alias{sigma.ZIPLNfit}
\title{Extract variance-covariance of residuals 'Sigma'}
\usage{
\method{sigma}{ZIPLNfit}(object, ...)
}
\arguments{
\item{object}{an R6 object with class \code{\link{ZIPLNfit}}}

\item{...}{additional parameters for S3 compatibility. Not used}
}
\value{
A semi definite positive matrix of size p, assuming there are p species in the model.
}
\description{
Extract the variance-covariance matrix of the residuals, usually noted \eqn{\Sigma} in ZIPLN models.
}
\seealso{
\code{\link[=coef.ZIPLNfit]{coef.ZIPLNfit()}}
}
