% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNfit-S3methods.R
\name{standard_error}
\alias{standard_error}
\title{Component-wise standard errors of Theta}
\usage{
standard_error(x, ...)
}
\arguments{
\item{x}{An `R` object. Currently there are methods for \code{\link{PLNfit}} (and its variants) objects.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
The standard errors associated with coefficients of \eqn{\Theta}
}
\description{
Compute or extract component-wise standard errors of Theta in multivariate (generalized) linear models of the form \deqn{g(E[Y|X]) = X\Theta} Useful to compute Z-scores and p-values under a gaussian/student approximation of \eqn{\Theta}
}
