% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNnetworkfamily-class.R
\name{stars}
\alias{stars}
\title{Perform StARS (Stability Approach to Regularization Selection)}
\arguments{
\item{stability}{a scalar, indicating the target stability (= 1 - 2 beta) at which the network is selected. Default is \code{0.9}.}

\item{subsamples}{a list of vectors describing the subsamples. The number of vectors (or list length) determines th number of subsamples used in the stability selection. Automatically set to 20 subsamples with size \code{10*sqrt(n)} if \code{n >= 144} and \code{0.8*n} otherwise following Liu et al. (2010) recommandations.}
}
\description{
Compute StARS criterion after stability selection. If a stability path is already present,
it will not be recomputed
}
