% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{starting_point}
\alias{starting_point}
\title{Barebone function to compute starting points for B, M and S}
\usage{
starting_point(Y, X, O, w)
}
\description{
Barebone function to compute starting points for B, M and S
}
