% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trichoptera.R
\docType{data}
\name{trichoptera}
\alias{trichoptera}
\title{Trichoptera data set}
\format{A data frame with 49 rows and 3 variables (matrices):
\describe{
  \item{abundance}{a matrix with 49 rows (trapping nights) and 17 species.}
  \item{meteo}{a matrix with 49 rows and 11 meteorological variables.}
  \item{group}{a factor of 12 levels for the definition of the consecutive night groups}
}
This format is convenient for using formula in multivariate analysis (multiple outputs and inputs).}
\source{
Data from P. Usseglio-Polatera.
}
\usage{
trichoptera
}
\description{
Data gathered between 1959 and 1960 during 49 insect trapping nights.
For each trapping night, the abundance of 17 Trichoptera
species is recorded, which forms the first table.
The second table concerns 11 meteorological variables which may influence
the abundance of each species. Finally, the observations (that is to say,
the trapping nights), have been classified into 12 groups corresponding to
contiguous nights between summer 1959 and summer 1960.
}
\examples{
data(trichoptera)
str(trichopetra)
## also see the package vignettes

}
\references{
Usseglio-Polatera, P. and Auda, Y. (1987) Influence des facteurs météorologiques sur les résultats de piégeage lumineux. Annales de Limnologie, 23, 65–79. (code des espèces p. 76)
See a data description at \href{http://pbil.univ-lyon1.fr/R/pdf/pps034.pdf}{http://pbil.univ-lyon1.fr/R/pdf/pps034.pdf} (in French)
}
\keyword{datasets}
