% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNARfit-S3methods.R
\name{vcov.PLNARfit}
\alias{vcov.PLNARfit}
\title{Calculate Variance-Covariance Matrix for a fitted \code{\link[=PLN]{PLN()}} model object}
\usage{
\method{vcov}{PLNARfit}(object, type = c("main", "covariance"), ...)
}
\arguments{
\item{object}{an R6 object with class \code{\link{PLNARfit}}}

\item{type}{type of parameter that should be extracted. Either "main" (default) for \deqn{B} or "covariance" for \deqn{\Sigma}}

\item{...}{additional parameters for S3 compatibility. Not used}
}
\value{
A matrix of variance/covariance extracted from the PLNARfit model. If type="main" and \eqn{B} is a matrix of size d * p, the result is a block-diagonal matrix with p (number of species) blocks of size d (number of covariates). if type="main", it is a symmetric matrix of size p.
.
}
\description{
Returns the variance-covariance matrix of the main parameters of a fitted \code{\link[=PLN]{PLN()}} model object. The main parameters of the model correspond to \deqn{B}, as returned by \code{\link[=coef.PLNARfit]{coef.PLNARfit()}}. The function can also be used to return the variance-covariance matrix of the residuals. The latter matrix can also be accessed via \code{\link[=sigma.PLNARfit]{sigma.PLNARfit()}}
}
\examples{
data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPLN <- PLN(Abundance ~ 1 + offset(log(Offset)), data = trichoptera)
vcov(myPLN, type = "covariance") ## Sigma
}
\seealso{
\code{\link[=sigma.PLNARfit]{sigma.PLNARfit()}}, \code{\link[=coef.PLNARfit]{coef.PLNARfit()}}, \code{\link[=standard_error.PLNARfit]{standard_error.PLNARfit()}}
}
