% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNfit-class.R
\name{vcov.PLNfit}
\alias{vcov.PLNfit}
\title{Extracts model covariance from objects returned by \code{\link[=PLN]{PLN}} and its variants}
\usage{
\method{vcov}{PLNfit}(object, ...)
}
\arguments{
\item{object}{an R6 object with class PLNfit}

\item{...}{additional parameters for S3 compatibility. Not used}
}
\value{
A matrix of variance/covariance extracted from the PLNfit model.
}
\description{
Extracts model covariance from objects returned by \code{\link[=PLN]{PLN}} and its variants
}
