% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIPLN.R
\name{ZIPLN}
\alias{ZIPLN}
\title{Zero Inflated Poisson lognormal model}
\usage{
ZIPLN(formula, data, subset, rho = 0, lambda = 0, control = ZIPLN_param())
}
\arguments{
\item{formula}{an object of class "formula": a symbolic description of the model to be fitted.}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which PLN is called.}

\item{subset}{an optional vector specifying a subset of observations to be used in the fitting process.}

\item{rho}{a positive number controlling the level of sparsity in the inverse covariance matrix. Default to 0 (no sparsity).}

\item{lambda}{a positive number controlling the level of sparsity in the regression parameters of the PLN component. Default to 0 (no sparsity).}

\item{control}{a list-like structure for controlling the optimization, with default generated by \code{\link[=ZIPLN_param]{ZIPLN_param()}}. See the associated documentation
for details.}
}
\value{
an R6 object with class \code{\link{ZIPLNfit}}
}
\description{
Fit the multivariate Zero Inflated Poisson lognormal model with a variational algorithm. Use the (g)lm syntax for model specification (covariates, offsets, subset).
}
\examples{
data(scRNA)
# data subsample: only 100 random cell and the 50 most varying transcript
scRNA        <- scRNA[sample.int(nrow(scRNA), 100), ]
scRNA$counts <- scRNA$counts[, 1:50]
myPLN_full   <- ZIPLN(counts ~ 1 + cell_line + offset(log(total_counts)) | 1, data = scRNA)
myPLN_sparse <- ZIPLN(counts ~ 1 + offset(log(total_counts)), rho = .5, data = scRNA)
myPLN_full$criteria    # better BIC with sparse version
myPLN_sparse$criteria
}
\seealso{
The class \code{\link{ZIPLNfit}}
}
