% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNfit-S3methods.R
\name{fisher.PLNfit}
\alias{fisher.PLNfit}
\title{Extracts (one-data) Fisher information matrix of Theta from objects returned by \code{\link[=PLN]{PLN}} and its variants}
\usage{
fisher.PLNfit(object, type = c("wald", "louis"), X = NULL, ...)
}
\arguments{
\item{object}{an R6 object with class PLNfit}

\item{type}{Either `Wald` (default) or `Louis`. Approxomation scheme used to compute the
Fisher information matrix}

\item{X}{Required. The covariate matrix used to fit the model.}

\item{...}{additional parameters for S3 compatibility. Not used}
}
\value{
A block-diagonal matrix with p (number of species) blocks of size d (number of covariates), assuming
Theta is a matrix of size d * p.
}
\description{
Extracts (one-data) Fisher information matrix of Theta from objects returned by \code{\link[=PLN]{PLN}} and its variants
}
