% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3methods.R
\name{getBestModel.PLNPCAfamily}
\alias{getBestModel.PLNPCAfamily}
\title{Best model extraction from a collection of PLNPCAfit}
\usage{
\method{getBestModel}{PLNPCAfamily}(Robject, crit = c("ICL", "BIC",
  "R_squared"), ...)
}
\arguments{
\item{Robject}{an object with classi PLNPCAfamilly}

\item{crit}{a character for the criterion used to performed the selection. Either
"BIC", "ICL", "R_squared". Default is \code{ICL}.}

\item{...}{not use.}
}
\value{
Send back a object with class \code{\link[=PLNPCAfit]{PLNPCAfit}}.
}
\description{
Best model extraction from a collection of PLNPCAfit
}
