% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLN.R
\name{PLN}
\alias{PLN}
\alias{PLN.formula}
\alias{PLN.default}
\title{Fit a Poisson lognormal model with a variational algorithm}
\usage{
PLN(Robject, ...)

\method{PLN}{formula}(Robject, control = list(), ...)

\method{PLN}{default}(Robject, X = NULL, O = NULL, control = list(),
  ...)
}
\arguments{
\item{Robject}{an R object, either a formula or a (n x p) matrix of count data}

\item{...}{additional parameters for S3 compatibility. Not used}

\item{control}{a list for controlling the optimization. See details.}

\item{X}{an optional (n x d) matrix of covariates. A vector of intercept is included by default. Ignored when Robject is a formula.}

\item{O}{an optional (n x p) matrix of offsets. Ignored when Robject is a formula.}
}
\value{
an R6 object with class \code{\link[=PLNfit]{PLNfit}}
}
\description{
two methods are available for specifying the models (with formulas or matrices)
}
\details{
The parameter \code{control} is a list controlling the optimization with the following entries
\itemize{
 \item{"ftol_rel"}{stop when an optimization step changes the objective function by less than ftol multiplied by the absolute value of the parameter. Default is 1e-6}
 \item{"ftol_abs"}{stop when an optimization step changes the objective function by less than ftol multiplied by the absolute value of the parameter. Default is 1e-6}
 \item{"xtol_rel"}{stop when an optimization step changes every parameters by less than xtol multiplied by the absolute value of the parameter. Default is 1e-4}
 \item{"xtol_abs"}{stop when an optimization step changes every parameters by less than xtol multiplied by the absolute value of the parameter. Default is 1e-4}
 \item{"maxeval"}{stop when the number of iteration exceeds maxeval. Default is 10000}
 \item{"method"}{the optimization method used by NLOPT among LD type, i.e. "MMA", "LBFGS",
    "TNEWTON", "TNEWTON_RESTART", "TNEWTON_PRECOND", "TNEWTON_PRECOND_RESTART",
    "TNEWTON_VAR1", "TNEWTON_VAR2". See NLOPTR documentation for further details. Default is "MMA".}
 \item{"lbvar"}{the lower bound (box constraint) for the variational variance parameters. Default is 1e-5.}
 \item{"trace"}{integer for verbosity. Useless when \code{cores} > 1}
 \item{"inception"}{How to setup the intialization: either with a PLNfit (typically obtained from a previsous fit), or a character string between "LM" and "GLM".
  If "LM", a log transformation is applied to Y then a linear model for initialization.
  If "GLM", a GLM Poisson is used. Default is "LM".}
}
}
\examples{
## See the vignette
}
\seealso{
The class  \code{\link[=PLNfit]{PLNfit}}
}
