% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLN.R
\name{PLN}
\alias{PLN}
\alias{PLN.formula}
\alias{PLN.default}
\title{Fit a Poisson lognormal model}
\usage{
PLN(Robject, ...)

\method{PLN}{formula}(formula, control = list())

\method{PLN}{default}(Y, X = cbind(rep(1, nrow(Y))), O = matrix(0, nrow(Y),
  ncol(Y)), control = list())
}
\arguments{
\item{Robject}{an R object, either a formula or a matrix}

\item{...}{additional parameters. Not used}

\item{formula}{a formula}

\item{control}{a list for controling the optimization. See details.}

\item{Y}{a (n x p) matrix of count data}

\item{X}{an optional (n x d) matrix of covariates. SHould include the intercept (a column of one) if the default method is used.}

\item{O}{an optional (n x p) matrix of offsets.}
}
\value{
an R6 object with class \code{\link[=PLNfit-class]{PLNfit}}
}
\description{
two methods are available for specifing the models (with formulas or matrices)
}
\details{
The parameter \code{control} is a list controling the optimization with the following entries
\itemize{
 \item{"xtol"}{stop when an optimization step changes every parameters by less than xtol multiply by the absolute value of the parameter. Default is 1e-4}
 \item{"ftol"}{stop when an optimization step changes the objective function by less than xtol multiply by the absolute value of the parameter. Default is 1e-6}
 \item{"maxit"}{stop when the number of iteration exeeeds maxiter. Default is 10000}
 \item{"lbvar"}{the lower bound (box constraint) for the variational variance parameters. Default is 1e-5.}
 \item{"trace"}{integer for verbosity. Useless when \code{cores} > 1}
}
}
\examples{
## See the vignette TODO!!!
}
\seealso{
The class  \code{\link[=PLNfit-class]{PLNfit}}
}
