% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNblockfit-class.R
\name{PLNblockfit}
\alias{PLNblockfit}
\title{An R6 Class to represent a PLNfit with block-wise residual covariance}
\description{
An R6 Class to represent a PLNfit with block-wise residual covariance

An R6 Class to represent a PLNfit with block-wise residual covariance
}
\examples{
\dontrun{
data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPLN <- PLNblock(Abundance ~ 1, data = trichoptera, nb_blocks = 1:15)
class(myPLN)
print(myPLN)
}
}
\section{Super class}{
\code{\link[PLNmodels:PLNfit]{PLNmodels::PLNfit}} -> \code{PLNblockfit}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{nb_param}}{number of parameters in the current PLN model}

\item{\code{nb_block}}{number blocks of variables (dimension of the residual covariance)}

\item{\code{vcov_model}}{character: the model used for the residual covariance}

\item{\code{loglik}}{(weighted) variational lower bound of the loglikelihood}

\item{\code{loglik_vec}}{element-wise variational lower bound of the loglikelihood}

\item{\code{entropy_blocks}}{Entropy of the variational distribution of the block (multinomial)}

\item{\code{entropy_latent}}{Entropy of the variational distribution of the latent vector (Gaussian)}

\item{\code{entropy}}{Full entropy of the variational distribution (latent vector + block)}

\item{\code{membership}}{block membership for variables}

\item{\code{posteriorProb}}{matrix of posterior probabilities for block belonging}

\item{\code{groupProportion}}{vector of cluster proportions}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PLNblockfit-new}{\code{PLNblockfit$new()}}
\item \href{#method-PLNblockfit-update}{\code{PLNblockfit$update()}}
\item \href{#method-PLNblockfit-optimize}{\code{PLNblockfit$optimize()}}
\item \href{#method-PLNblockfit-clone}{\code{PLNblockfit$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="optimize_vestep"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-optimize_vestep'><code>PLNmodels::PLNfit$optimize_vestep()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="postTreatment"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-postTreatment'><code>PLNmodels::PLNfit$postTreatment()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="predict"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-predict'><code>PLNmodels::PLNfit$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="predict_cond"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-predict_cond'><code>PLNmodels::PLNfit$predict_cond()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="print"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-print'><code>PLNmodels::PLNfit$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfit" data-id="show"><a href='../../PLNmodels/html/PLNfit.html#method-PLNfit-show'><code>PLNmodels::PLNfit$show()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PLNblockfit-new"></a>}}
\if{latex}{\out{\hypertarget{method-PLNblockfit-new}{}}}
\subsection{Method \code{new()}}{
Initialize a \code{\link{PLNblockfit}} model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PLNblockfit$new(
  blocks,
  responses,
  covariates,
  offsets,
  weights,
  formula,
  control
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{blocks}}{blocks indicators for regrouping the variables/responses (dimension of the residual covariance)}

\item{\code{responses}}{the matrix of responses (called Y in the model). Will usually be extracted from the corresponding field in PLNfamily-class}

\item{\code{covariates}}{design matrix (called X in the model). Will usually be extracted from the corresponding field in PLNfamily-class}

\item{\code{offsets}}{offset matrix (called O in the model). Will usually be extracted from the corresponding field in PLNfamily-class}

\item{\code{weights}}{an optional vector of observation weights to be used in the fitting process.}

\item{\code{formula}}{model formula used for fitting, extracted from the formula in the upper-level call}

\item{\code{control}}{a list for controlling the optimization.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PLNblockfit-update"></a>}}
\if{latex}{\out{\hypertarget{method-PLNblockfit-update}{}}}
\subsection{Method \code{update()}}{
Update fields of a \code{\link{PLNnetworkfit}} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PLNblockfit$update(
  Tau = NA,
  B = NA,
  Sigma = NA,
  Omega = NA,
  M = NA,
  S = NA,
  Z = NA,
  A = NA,
  Ji = NA,
  R2 = NA,
  monitoring = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Tau}}{matrix of posterior probabilities for block belonging}

\item{\code{B}}{matrix of regression matrix}

\item{\code{Sigma}}{variance-covariance matrix of the latent variables}

\item{\code{Omega}}{precision matrix of the latent variables. Inverse of Sigma.}

\item{\code{M}}{matrix of mean vectors for the variational approximation}

\item{\code{S}}{matrix of variance vectors for the variational approximation}

\item{\code{Z}}{matrix of latent vectors (includes covariates and offset effects)}

\item{\code{A}}{matrix of fitted values}

\item{\code{Ji}}{vector of variational lower bounds of the log-likelihoods (one value per sample)}

\item{\code{R2}}{approximate R^2 goodness-of-fit criterion}

\item{\code{monitoring}}{a list with optimization monitoring quantities}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PLNblockfit-optimize"></a>}}
\if{latex}{\out{\hypertarget{method-PLNblockfit-optimize}{}}}
\subsection{Method \code{optimize()}}{
Call to the NLopt or TORCH optimizer and update of the relevant fields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PLNblockfit$optimize(responses, covariates, offsets, weights, config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{responses}}{the matrix of responses (called Y in the model). Will usually be extracted from the corresponding field in PLNfamily-class}

\item{\code{covariates}}{design matrix (called X in the model). Will usually be extracted from the corresponding field in PLNfamily-class}

\item{\code{offsets}}{offset matrix (called O in the model). Will usually be extracted from the corresponding field in PLNfamily-class}

\item{\code{weights}}{an optional vector of observation weights to be used in the fitting process.}

\item{\code{config}}{part of the \code{control} argument which configures the optimizer}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PLNblockfit-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PLNblockfit-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PLNblockfit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
