% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNfamily-class.R
\name{PLNfamily_getBestModel}
\alias{PLNfamily_getBestModel}
\title{Best model extraction from a collection of PLNfit (PCA, network)}
\arguments{
\item{crit}{a character for the criterion used to performed the selection. Either
"BIC", "ICL", "J", "R2". Default is "BIC".}
}
\value{
Send back a object with class \code{\link[=PLNfit]{PLNfit}}.
}
\description{
Best model extraction from a collection of PLNfit (PCA, network)
}
