% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNfamily-class.R
\name{PLNfamily_getModel}
\alias{PLNfamily_getModel}
\title{Model extraction from a collection of PLNPCAfit}
\arguments{
\item{xvar}{a number (rank for PCA, penalty for network) identifying the model to be extracted from the collection.}
}
\value{
Send back a object with class \code{\link[=PLNfit]{PLNfit}}.
}
\description{
Model extraction from a collection of PLNPCAfit
}
