% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNfamily-class.R
\name{PLNfamily_getModel}
\alias{PLNfamily_getModel}
\title{Model extraction from a collection of PLN models}
\arguments{
\item{var}{value of the parameter (rank for PCA, penalty for network) that identifies the model to be extracted from the collection.}
}
\value{
Send back a object with class \code{\link[=PLNfit]{PLNfit}}.
}
\description{
Model extraction from a collection of PLN models
}
