% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNfamily-class.R
\name{PLNfamily_getModel}
\alias{PLNfamily_getModel}
\title{Model extraction from a collection of PLNfit}
\arguments{
\item{rank}{an integer given the rank of the model to be extracted from the collection.}
}
\value{
Send back a object with class \code{\link[=PLNfit.PCA-class]{PLNfit}}.
}
\description{
Model extraction from a collection of PLNfit
}

