% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNfit-class.R
\docType{data}
\name{PLNfit}
\alias{PLNfit}
\title{An R6 Class to represent a PLNfit in a standard, general framework}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PLNfit
}
\description{
The function \code{\link{PLN}} produces a collection of models which are instances of object with class \code{PLNfit}.
Objects produced by the functions \code{\link{PLNnetwork}} and \code{\link{PLNPCA}} also enjoy the method of \code{\link{PLNfit}}
by inheritance.

This class comes with a set of methods, some of them being useful for the user: plot_model_par, plot_variational_par

Fields should not be changed or manipulated by the user as they are updated internally.
}
\section{Fields}{

\describe{
\item{\code{model_par}}{a list with two matrices, B and Theta, which are the estimated parameters of the pPCA model}

\item{\code{var_par}}{a list with two matrices, M and S, which are the estimated parameters in the variational approximation}

\item{\code{criteria}}{a named vector with the value of some criteria (variational lower bound J, BIC, ICL, R2, lmin and lmax) for the different models.}

\item{\code{convergence}}{quantities useful for monitoring the optimization}
}}

\keyword{datasets}
