% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNfit-class.R
\name{PLNfit_VEstep}
\alias{PLNfit_VEstep}
\title{Result of the VE step of the optimization procedure: optimal variational parameters (M, S)
and corresponding log likelihood values of new observations for fixed model parameters (Sigma, Theta)}
\arguments{
\item{X}{A matrix of covariates.}

\item{O}{A matrix of offsets.}

\item{Y}{A matrix of counts.}

\item{control}{a list for controlling the optimization. See \code{\link[=PLN]{PLN}} for details.}
}
\value{
A list with three components:
           the matrix M of variational means,
           the matrix S of variational variances
           the vector log.lik of (variational) log-likelihood of each new observation
}
\description{
Result of the VE step of the optimization procedure: optimal variational parameters (M, S)
and corresponding log likelihood values of new observations for fixed model parameters (Sigma, Theta)
}
