% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNLDA.R
\name{PLNLDA}
\alias{PLNLDA}
\alias{PLNLDA.formula}
\alias{PLNLDA.default}
\title{Fit a Poisson lognormal model towards LInear Disciminant Analysis}
\usage{
PLNLDA(Robject, ...)

\method{PLNLDA}{formula}(Robject, grouping, control, ...)

\method{PLNLDA}{default}(Robject, grouping, X = NULL, O = matrix(0, nrow(Y),
  ncol(Y)), control = list(), ...)
}
\arguments{
\item{Robject}{either (n x p) matrix of count data (used with argument grouping) or a formula to describe the relationship between the count matrix and the covariates (apart from the grouping)}

\item{...}{additional parameters. Not used}

\item{grouping}{a factor specifying the class fo< each observation used for disciminant analysis. Ignored when Robject is a formula.}

\item{control}{a list for controling the optimization process. See details.}

\item{X}{an optional (n x d) matrix of covariates. A vector of intercept is included by default. Ignored when Robject is a formula.}

\item{O}{an optional (n x p) matrix of offsets. Ignored when Robject is a formula.}
}
\value{
an R6 object with class \code{\link[=PLNLDAfit]{PLNLDAfit}}
}
\description{
two methods are available for specifing the models (with formulas or matrices)
}
\details{
The list of parameters \code{control.init} and \code{control.main} control the optimization of the intialization and the main process, with the following entries
\itemize{
  \item{"ftol_rel"}{stop when an optimization step changes the objective function by less than ftol_rel multiplied by the absolute value of the parameter.}
 \item{"ftol_abs"}{stop when an optimization step changes the objective function by less than ftol_abs .}
 \item{"xtol_rel"}{stop when an optimization step changes every parameters by less than xtol_rel multiplied by the absolute value of the parameter.}
 \item{"xtol_abs"}{stop when an optimization step changes every parameters by less than xtol_abs.}
 \item{"maxeval"}{stop when the number of iteration exceeds maxeval. Default is 10000}
 \item{"method"}{the optimization method used by NLOPT among LD type, i.e. "MMA", "LBFGS",
    "TNEWTON", "TNEWTON_RESTART", "TNEWTON_PRECOND", "TNEWTON_PRECOND_RESTART",
    "TNEWTON_VAR1", "TNEWTON_VAR2". See NLOPTR documentation for further details. Default is "MMA".}
 \item{"lbvar"}{the lower bound (box constraint) for the variational variance parameters. Default is 1e-5.}
 \item{"trace"}{integer for verbosity. Useless when \code{cores} > 1}
 \item{"inception"}{a PLNfit to start with. If NULL, a PLN is fitted on the . If an R6 object with class 'PLNfit' is given, it is used to initialize the model.}
 \item{"cores"}{The number of core used to paralellize jobs over the \code{ranks} vector. Default is 1.}
}
}
\examples{
## See the vignette: vignette("trichoptera", package="PLNmodels")
}
\seealso{
The class \code{\link[=PLNPCAfit]{PLNLDAfit}}
}
