% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNMMfamily-class.R
\docType{data}
\name{PLNMMfamily}
\alias{PLNMMfamily}
\title{An R6 Class to represent a collection of PLNMMfit}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PLNMMfamily
}
\description{
The function \code{\link{PLNMM}} produces an instance of this class.

This class comes with a set of methods, some of them being useful for the user:
See the documentation for \code{\link[=PLNfamily_getBestModel]{getBestModel}},
\code{\link[=PLNfamily_getModel]{getModel}}, \code{\link[=plot.PLNfamily]{plot}}
and \code{\link[=predict.PLNfit]{predict}}.
}
\section{Fields}{

\describe{
\item{\code{responses}}{the matrix of responses common to every models}

\item{\code{covariates}}{the matrix of covariates common to every models}

\item{\code{offsets}}{the matrix of offsets common to every models}

\item{\code{clusters}}{the dimensions of the successively fitted models}

\item{\code{models}}{a list of \code{\link[=PLNMMfit]{PLNMMfit}} object, one for each number of clusters considered.}

\item{\code{inception}}{a \code{\link[=PLNfit]{PLNfit}} object, obtained when one single class is considered.}

\item{\code{criteria}}{a data frame with the value of some criteria (variational lower bound J, BIC, ICL and R2) for the different models.}
}}

\seealso{
The function \code{\link{PLNMM}}, the class \code{\link[=PLNMMfit]{PLNMMfit}}
}
\keyword{datasets}
