% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNMMfit-class.R
\name{PLNMMfit}
\alias{PLNMMfit}
\title{An R6 Class to represent a PLNfit in a mixture framework}
\description{
The function \code{\link{PLNMM}} produces a collection of models which are instances of object with class \code{PLNMMfit}.
A \code{PLNMMfit} (say, with k components) is itself a collection of k \code{PLNfit}.

This class comes with a set of methods, some of them being useful for the user:
See the documentation for ...
}
\seealso{
The function \code{\link{PLNMM}}, the class \code{\link[=PLNMMfamily]{PLNMMfamily}}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{meberships}}{vector for cluster index}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{n}}{number of samples}

\item{\code{k}}{number of components}

\item{\code{components}}{components of the mixture (PLNfits)}

\item{\code{posteriorProb}}{matrix ofposterior probability for cluster belonging}

\item{\code{meberships}}{vector for cluster index}

\item{\code{mixtureParam}}{vector of cluster proporitions}

\item{\code{nb_param}}{number of parameters in the current PLN model}

\item{\code{entropy}}{Entropy of the variational distribution of the cluster (multinomial)}

\item{\code{loglik}}{variational lower bound of the loglikelihood}

\item{\code{loglik_vec}}{element-wise variational lower bound of the loglikelihood}

\item{\code{R_squared}}{approximated goodness-of-fit criterion}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PLNMMfit$new()}}
\item \href{#method-update}{\code{PLNMMfit$update()}}
\item \href{#method-optimize}{\code{PLNMMfit$optimize()}}
\item \href{#method-postTreatment}{\code{PLNMMfit$postTreatment()}}
\item \href{#method-show}{\code{PLNMMfit$show()}}
\item \href{#method-print}{\code{PLNMMfit$print()}}
\item \href{#method-clone}{\code{PLNMMfit$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PLNMMfit$new(inception, tau, monitoring = NA, R2 = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update"></a>}}
\if{latex}{\out{\hypertarget{method-update}{}}}
\subsection{Method \code{update()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PLNMMfit$update(tau = NA, J = NA, R2 = NA, monitoring = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-optimize"></a>}}
\if{latex}{\out{\hypertarget{method-optimize}{}}}
\subsection{Method \code{optimize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PLNMMfit$optimize(responses, covariates, offsets, tau, opts)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-postTreatment"></a>}}
\if{latex}{\out{\hypertarget{method-postTreatment}{}}}
\subsection{Method \code{postTreatment()}}{
Update fields after optimization
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PLNMMfit$postTreatment(responses, covariates, offsets, weights, nullModel)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-show"></a>}}
\if{latex}{\out{\hypertarget{method-show}{}}}
\subsection{Method \code{show()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PLNMMfit$show()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PLNMMfit$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PLNMMfit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
