% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNMMfit-class.R
\docType{data}
\name{PLNMMfit}
\alias{PLNMMfit}
\title{An R6 Class to represent a PLNfit in a mixture framework}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PLNMMfit
}
\description{
The function \code{\link{PLNMM}} produces a collection of models which are instances of object with class \code{PLNMMfit}.
A \code{PLNMMfit} (say, with k components) is itself a collection of k \code{PLNfit}.

This class comes with a set of methods, some of them being useful for the user:
See the documentation for ...
}
\section{Fields}{

\describe{
\item{\code{cluster}}{the number of clusters of the current model}

\item{\code{components}}{a list with cluster component element, each of whom is a \code{PLNfit}.}

\item{\code{model_par}}{a list with the matrices associated with the final estimated parameters of the mixture model: Theta (covariates), Sigma (latent covariance), mu (vector of means/centers) and pi (vector of cluster proportions)}

\item{\code{posteriorProbabilities}}{matrix of posterior probabilities of class belonging}

\item{\code{mixtureParam}}{vector of cluster proportions}

\item{\code{loglik}}{variational lower bound of the loglikelihood}

\item{\code{BIC}}{variational lower bound of the BIC}

\item{\code{ICL}}{variational lower bound of the ICL}

\item{\code{R_squared}}{approximated goodness-of-fit criterion}

\item{\code{criteria}}{a vector with loglik, BIC, ICL, R_squared and degrees of freedom}

\item{\code{degrees_freedom}}{number of parameters in the current model}
}}

\seealso{
The function \code{\link{PLNMM}}, the class \code{\link[=PLNMMfamily]{PLNMMfamily}}
}
\keyword{datasets}
