% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNnetwork.R
\name{PLNnetwork}
\alias{PLNnetwork}
\alias{PLNnetwork.formula}
\alias{PLNnetwork.default}
\title{Fit a Poisson lognormal model towards network inference}
\usage{
PLNnetwork(Robject, ...)

\method{PLNnetwork}{formula}(formula, penalties = NULL, control = list())

\method{PLNnetwork}{default}(Y, X = cbind(rep(1, nrow(Y))), O = matrix(0,
  nrow(Y), ncol(Y)), penalties = NULL, control = list())
}
\arguments{
\item{Robject}{an R object, either a formula or a matrix}

\item{...}{additional parameters. Not used}

\item{formula}{a formula}

\item{penalties}{a vector of positive real number controling the level of sparisty of the underlying network}

\item{control}{a list for controling the optimization. See details.}

\item{Y}{a (n x p) matrix of count data}

\item{X}{an optional (n x d) matrix of covariates. SHould include the intercept (a column of one) if the default method is used.}

\item{O}{an optional (n x p) matrix of offsets.}
}
\value{
an R6 object with class \code{\link[=PLNnetworkfamily-class]{PLNnetworkfamily}}, which contains
a collection of models with class \code{\link[=PLNnetworkfit-class]{PLNnetworkfit}}
}
\description{
two methods are available for specifing the models (with formulas or matrices)
}
\details{
The parameter \code{control} is a list with the following entries
\itemize{
 \item{"out.tol"}{outer solver stops when an optimization step changes the objective function by less than xtol multiply by the absolute value of the parameter. Default is 1e-6}
 \item{"out.maxit"}{outer solver stops when the number of iteration exeeeds maxiter. Default is 10000}
 \item{"xtol"}{inner solver stops when an optimization step changes every parameters by less than xtol multiply by the absolute value of the parameter. Default is 1e-4}
 \item{"ftol"}{inner solver stops when an optimization step changes the objective function by less than xtol multiply by the absolute value of the parameter. Default is 1e-6}
 \item{"maxit"}{inner solver stops when the number of iteration exeeeds maxiter. Default is 10000}
 \item{"lbvar"}{the lower bound (box constraint) for the variational variance parameters. Default is .Machine$double.eps.}
 \item{"lbvar.init"}{the lower bound (box constraint) for the variational variance parameters for the unpenalized model. Default is 1e-5.}
 \item{"trace"}{integer for verbosity. Useless when \code{cores} > 1}
}
}
\examples{
## See the vignette TODO
}
\seealso{
The classes \code{\link[=PLNnetworkfamily-class]{PLNnetworkfamily}} and \code{\link[=PLNnetworkfit-class]{PLNnetworkfit}}
}
