% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNnetworkfamily-class.R
\docType{data}
\name{PLNnetworkfamily}
\alias{PLNnetworkfamily}
\title{An R6 Class to represent a collection of PLNnetworkfit}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PLNnetworkfamily
}
\description{
The function \code{\link{PLNnetwork}} produces an instance of this class.

This class comes with a set of methods, some of them being useful for the user:
See the documentation for \code{\link[=PLNnetworkfamily_getBestModel]{getBestModel}},
\code{\link[=PLNnetworkfamily_getModel]{getModel}} and \code{\link[=PLNnetworkfamily_plot]{plot}}. Other methods
 should not be called as they are designed to be used during the optimization process.

Fields should not be changed or manipulated by the user as they are updated internally
during the estimation process.
}
\section{Fields}{

\describe{
\item{\code{penalties}}{the sparsity level of the network in the successively fitted models}

\item{\code{models}}{a list of \code{\link[=PLNnetworkfit-class]{PLNnetworkfit}} object, one per penalty.}

\item{\code{model.unpenalized}}{a \code{\link[=PLNnetworkfit-class]{PLNnetworkfit}} object, obtained when no sparsifying penalty is applied.}

\item{\code{criteria}}{a data frame with the value of some criteria (variational lower bound J, BIC, ICL and R2) for the different models.}

\item{\code{responses}}{the matrix of responses common to every models}

\item{\code{covariates}}{the matrix of covariates common to every models}

\item{\code{offsets}}{the matrix of offsets common to every models}

\item{\code{objective}}{the R function used to compute the model's objective during the optimization process}

\item{\code{gradient}}{the R function to compute the model's gradient during the optimization process}
}}

\seealso{
The function \code{\link{PLNnetwork}}, the class \code{\link[=PLNnetworkfit-class]{PLNnetworkfit}}
}
\keyword{datasets}
