% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNnetworkfamily-class.R
\name{PLNnetworkfamily}
\alias{PLNnetworkfamily}
\title{An R6 Class to represent a collection of PLNnetworkfit}
\description{
The function \code{\link[=PLNnetwork]{PLNnetwork()}} produces an instance of this class.

This class comes with a set of methods, some of them being useful for the user:
See the documentation for \code{\link[=getBestModel]{getBestModel()}},
\code{\link[=getModel]{getModel()}} and \link[=plot.PLNnetworkfamily]{plot()}
}
\examples{
data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
fits <- PLNnetwork(Abundance ~ 1, data = trichoptera)
class(fits)
}
\seealso{
The function \code{\link[=PLNnetwork]{PLNnetwork()}}, the class \code{\link{PLNnetworkfit}}
}
\section{Super classes}{
\code{\link[PLNmodels:PLNfamily]{PLNmodels::PLNfamily}} -> \code{PLNmodels::PLNnetworkfamilyvirtual} -> \code{PLNnetworkfamily}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PLNnetworkfamily-new}{\code{PLNnetworkfamily$new()}}
\item \href{#method-PLNnetworkfamily-stability_selection}{\code{PLNnetworkfamily$stability_selection()}}
\item \href{#method-PLNnetworkfamily-clone}{\code{PLNnetworkfamily$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfamily" data-id="getModel"><a href='../../PLNmodels/html/PLNfamily.html#method-PLNfamily-getModel'><code>PLNmodels::PLNfamily$getModel()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfamily" data-id="postTreatment"><a href='../../PLNmodels/html/PLNfamily.html#method-PLNfamily-postTreatment'><code>PLNmodels::PLNfamily$postTreatment()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNfamily" data-id="print"><a href='../../PLNmodels/html/PLNfamily.html#method-PLNfamily-print'><code>PLNmodels::PLNfamily$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNnetworkfamilyvirtual" data-id="coefficient_path"><a href='../../PLNmodels/html/PLNnetworkfamilyvirtual.html#method-PLNnetworkfamilyvirtual-coefficient_path'><code>PLNmodels::PLNnetworkfamilyvirtual$coefficient_path()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNnetworkfamilyvirtual" data-id="getBestModel"><a href='../../PLNmodels/html/PLNnetworkfamilyvirtual.html#method-PLNnetworkfamilyvirtual-getBestModel'><code>PLNmodels::PLNnetworkfamilyvirtual$getBestModel()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNnetworkfamilyvirtual" data-id="optimize"><a href='../../PLNmodels/html/PLNnetworkfamilyvirtual.html#method-PLNnetworkfamilyvirtual-optimize'><code>PLNmodels::PLNnetworkfamilyvirtual$optimize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNnetworkfamilyvirtual" data-id="plot"><a href='../../PLNmodels/html/PLNnetworkfamilyvirtual.html#method-PLNnetworkfamilyvirtual-plot'><code>PLNmodels::PLNnetworkfamilyvirtual$plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNnetworkfamilyvirtual" data-id="plot_objective"><a href='../../PLNmodels/html/PLNnetworkfamilyvirtual.html#method-PLNnetworkfamilyvirtual-plot_objective'><code>PLNmodels::PLNnetworkfamilyvirtual$plot_objective()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNnetworkfamilyvirtual" data-id="plot_stars"><a href='../../PLNmodels/html/PLNnetworkfamilyvirtual.html#method-PLNnetworkfamilyvirtual-plot_stars'><code>PLNmodels::PLNnetworkfamilyvirtual$plot_stars()</code></a></span></li>
<li><span class="pkg-link" data-pkg="PLNmodels" data-topic="PLNnetworkfamilyvirtual" data-id="show"><a href='../../PLNmodels/html/PLNnetworkfamilyvirtual.html#method-PLNnetworkfamilyvirtual-show'><code>PLNmodels::PLNnetworkfamilyvirtual$show()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PLNnetworkfamily-new"></a>}}
\if{latex}{\out{\hypertarget{method-PLNnetworkfamily-new}{}}}
\subsection{Method \code{new()}}{
Initialize all models in the collection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PLNnetworkfamily$new(
  penalties,
  responses,
  covariates,
  offsets,
  weights,
  formula,
  control
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{penalties}}{a vector of positive real number controlling the level of sparsity of the underlying network.}

\item{\code{responses}}{the matrix of responses common to every models}

\item{\code{covariates}}{the matrix of covariates common to every models}

\item{\code{offsets}}{the matrix of offsets common to every models}

\item{\code{weights}}{the vector of observation weights}

\item{\code{formula}}{model formula used for fitting, extracted from the formula in the upper-level call}

\item{\code{control}}{a list for controlling the optimization.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Update current \code{\link{PLNnetworkfit}} with smart starting values
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PLNnetworkfamily-stability_selection"></a>}}
\if{latex}{\out{\hypertarget{method-PLNnetworkfamily-stability_selection}{}}}
\subsection{Method \code{stability_selection()}}{
Compute the stability path by stability selection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PLNnetworkfamily$stability_selection(
  subsamples = NULL,
  control = PLNnetwork_param()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{subsamples}}{a list of vectors describing the subsamples. The number of vectors (or list length) determines the number of subsamples used in the stability selection. Automatically set to 20 subsamples with size \code{10*sqrt(n)} if \code{n >= 144} and \code{0.8*n} otherwise following Liu et al. (2010) recommendations.}

\item{\code{control}}{a list controlling the main optimization process in each call to PLNnetwork. See \code{\link[=PLNnetwork]{PLNnetwork()}} for details.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PLNnetworkfamily-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PLNnetworkfamily-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PLNnetworkfamily$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
