% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNnetworkfit-class.R
\docType{data}
\name{PLNnetworkfit}
\alias{PLNnetworkfit}
\title{An R6 Class to represent a PLNfit in a network inference framework}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PLNnetworkfit
}
\description{
The function \code{\link{PLNnetwork}} produces a collection of models which are instances of object with class \code{PLNnetworkfit}.

This class (will) come with a set of methods, some of them being useful for the user

Fields should not be changed or manipulated by the user as they are updated internally.
}
\section{Fields}{

\describe{
\item{\code{type}}{a character indicating the model used for the covariance matrix in the variational Gaussian approximation. Either "diagonal" or "spherical".}

\item{\code{penalty}}{the level of sparsity in the current model}

\item{\code{model.par}}{a list with two matrices, B and Theta, which are the estimated parameters of the pPCA model}

\item{\code{variation.par}}{a list with two matrices, M and S, which are the estimated parameters in the variational approximation}

\item{\code{criteria}}{a named vector with the value of some criteria (variational lower bound J, BIC, ICL, R2, lmin and lmax) for the different models.}

\item{\code{convergence}}{numeric; the convergence status of the L-BFGS-B method}

\item{\code{loglik}}{numeric; the Poisson loglikelihood of the current model}
}}

\seealso{
The function \code{\link{PLNnetwork}}, the class \code{\link[=PLNnetworkfamily]{PLNnetworkfamily}}
}
\keyword{datasets}
