% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNnetworkfit-class.R
\docType{data}
\name{PLNnetworkfit}
\alias{PLNnetworkfit}
\title{An R6 Class to represent a PLNfit in a network inference framework}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
PLNnetworkfit
}
\description{
The function \code{\link{PLNnetwork}} produces a collection of models which are instances of object with class \code{PLNnetworkfit}.

This class comes with a set of methods, some of them being useful for the user: plot_network + methods inhererited from PLNfit.
}
\section{Fields}{

\describe{
\item{\code{penalty}}{the level of sparsity in the current model}

\item{\code{model_par}}{a list with the matrices associated with the estimated parameters of the pPCA model: Theta (covariates), Sigma (latent covariance) and Theta (latent precision matrix)}

\item{\code{var_par}}{a list with two matrices, M and S, which are the estimated parameters in the variational approximation}

\item{\code{criteria}}{a named vector with the value of some criteria (variational lower bound J, BIC, ICL, R2) for the different models.}

\item{\code{convergence}}{quantities usefull for monitoring the optimization}
}}

\seealso{
The function \code{\link{PLNnetwork}}, the class \code{\link[=PLNnetworkfamily]{PLNnetworkfamily}}
}
\keyword{datasets}
