% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNPCA.R
\name{PLNPCA}
\alias{PLNPCA}
\alias{PLNPCA.formula}
\alias{PLNPCA.default}
\title{Fit a Poisson lognormal model towards Principal Component Analysis}
\usage{
PLNPCA(Robject, ...)

\method{PLNPCA}{formula}(formula, ranks = 1:5, control.init = list(),
  control.main = list())

\method{PLNPCA}{default}(Y, X = cbind(rep(1, nrow(Y))), O = matrix(0,
  nrow(Y), ncol(Y)), ranks = 1:5, control.init = list(),
  control.main = list())
}
\arguments{
\item{Robject}{an R object, either a formula or a matrix}

\item{...}{additional parameters. Not used}

\item{formula}{a formula}

\item{ranks}{a vector of integer containing the successive ranks (or number of axes to be considered)}

\item{Y}{a (n x p) matrix of count data}

\item{X}{an optional (n x d) matrix of covariates. Should include the intercept (a column of one) if the default method is used.}

\item{O}{an optional (n x p) matrix of offsets.}

\item{control}{a list for controlling the optimization. See details.}
}
\value{
an R6 object with class \code{\link[=PLNPCAfamily]{PLNPCAfamily}}, which contains
a collection of models with class \code{\link[=PLNPCAfit]{PLPCAfit}}
}
\description{
two methods are available for specifing the models (with formulas or matrices)
}
\examples{
## See the vignette: vignette("trichoptera", package="PLNmodels")
}
\seealso{
The classes \code{\link[=PLNnetworkfamily]{PLNPCAfamily}} and \code{\link[=PLNPCAfit]{PLNPCAfit}}
}
