% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNPCA.R
\name{PLNPCA}
\alias{PLNPCA}
\alias{PLNPCA.formula}
\alias{PLNPCA.default}
\title{Fit a Poisson lognormal model towards Principal Component Analysis}
\usage{
PLNPCA(Robject, ...)

\method{PLNPCA}{formula}(Robject, ranks = 1:5, control.init = list(),
  control.main = list())

\method{PLNPCA}{default}(Robject, X = matrix(1, nrow = nrow(Robject)),
  O = matrix(0, nrow(Robject), ncol(Robject)), ranks = 1:5,
  control.init = list(), control.main = list())
}
\arguments{
\item{Robject}{an R object, either a formula or a matrix}

\item{...}{additional parameters. Not used}

\item{ranks}{a vector of integer containing the successive ranks (or number of axes to be considered)}

\item{control.init}{a list for controling the optimization. See details.}

\item{control.main}{a list for controling the optimization. See details.}

\item{X}{an optional (n x d) matrix of covariates. Should include the intercept (a column of one) if the default method is used.}

\item{O}{an optional (n x p) matrix of offsets.}

\item{formula}{a formula}

\item{Y}{a (n x p) matrix of count data}
}
\value{
an R6 object with class \code{\link[=PLNPCAfamily]{PLNPCAfamily}}, which contains
a collection of models with class \code{\link[=PLNPCAfit]{PLPCAfit}}
}
\description{
two methods are available for specifing the models (with formulas or matrices)
}
\details{
The list of parameters \code{control.init} and \code{control.main} control the optimization of the intialization and the main process, with the following entries
\itemize{
  \item{"ftol_rel"}{stop when an optimization step changes the objective function by less than ftol_rel multiplied by the absolute value of the parameter.}
 \item{"ftol_abs"}{stop when an optimization step changes the objective function by less than ftol_abs .}
 \item{"xtol_rel"}{stop when an optimization step changes every parameters by less than xtol_rel multiplied by the absolute value of the parameter.}
 \item{"xtol_abs"}{stop when an optimization step changes every parameters by less than xtol_abs.}
 \item{"maxeval"}{stop when the number of iteration exceeds maxeval. Default is 10000}
 \item{"method"}{the optimization method used by NLOPT among LD type, i.e. "MMA", "LBFGS",
    "TNEWTON", "TNEWTON_RESTART", "TNEWTON_PRECOND", "TNEWTON_PRECOND_RESTART",
    "TNEWTON_VAR1", "TNEWTON_VAR2". See NLOPTR documentation for further details. Default is "MMA".}
 \item{"lbvar"}{the lower bound (box constraint) for the variational variance parameters. Default is 1e-5.}
 \item{"trace"}{integer for verbosity. Useless when \code{cores} > 1}
 \item{"inception"}{a PLNfit to start with. If NULL, a PLN is fitted on the . If an R6 object with class 'PLNfit' is given, it is used to initialize the model.}
 \item{"cores"}{The number of core used to paralellize jobs over the \code{ranks} vector. Default is 1.}
}
}
\examples{
## See the vignette: vignette("trichoptera", package="PLNmodels")
}
\seealso{
The classes \code{\link[=PLNnetworkfamily]{PLNPCAfamily}} and \code{\link[=PLNPCAfit]{PLNPCAfit}}
}
