% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNfit-S3methods.R
\name{standard_error.PLNfit}
\alias{standard_error.PLNfit}
\title{Component-wise standard errors of Theta}
\usage{
\method{standard_error}{PLNfit}(object, type = c("wald", "louis"), ...)
}
\arguments{
\item{object}{an R6 object with class PLNfit}

\item{type}{Either `Wald` (default) or `Louis`. Approxomation scheme used to compute the
Fisher information matrix}

\item{...}{additional parameters for S3 compatibility. Not used}
}
\value{
A p * d positive matrix (same size as \eqn{\Theta}) with standard errors for the coefficients of \eqn{\Theta}
}
\description{
Extracts univariate standard errors for the estimated coefficient of Theta. Standard errors are computed from the (approximate) Fisher information matrix. See \code{\link[=fisher.PLNfit]{fisher}} for more details on the approximations.
}
\seealso{
\code{\link[=fisher.PLNfit]{fisher}} for the complete Fisher information matrix
}
