% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLNfit-S3methods.R
\name{standard_error}
\alias{standard_error}
\alias{standard_error.PLNfit}
\alias{standard_error.PLNfit_fixedcov}
\title{Component-wise standard errors of Theta}
\usage{
standard_error(
  object,
  type = c("variational", "jackknife", "sandwich"),
  parameter = c("Theta", "Omega")
)

\method{standard_error}{PLNfit}(
  object,
  type = c("variational", "jackknife", "sandwich"),
  parameter = c("Theta", "Omega")
)

\method{standard_error}{PLNfit_fixedcov}(
  object,
  type = c("variational", "jackknife", "sandwich"),
  parameter = c("Theta", "Omega")
)
}
\arguments{
\item{object}{an R6 object with class PLNfit}

\item{type}{string describing the type of variance approximation: "variational", "jackknife", "sandwich" (only for fixed covariance). Default is "variational".}

\item{parameter}{string describing the target parameter: either Theta (regression coefficicents) or Omega (inverse residual covariance)}
}
\value{
A p * d positive matrix (same size as \eqn{\Theta}) with standard errors for the coefficients of \eqn{\Theta}
}
\description{
Extracts univariate standard errors for the estimated coefficient of Theta. Standard errors are computed from the (approximate) Fisher information matrix.
}
\section{Methods (by class)}{
\itemize{
\item \code{standard_error(PLNfit)}: Component-wise standard errors of Theta in \code{\link{PLNfit}}

\item \code{standard_error(PLNfit_fixedcov)}: Component-wise standard errors of Theta in \code{\link{PLNfit_fixedcov}}

}}
\examples{
data(trichoptera)
trichoptera <- prepare_data(trichoptera$Abundance, trichoptera$Covariate)
myPLN <- PLN(Abundance ~ 1 + offset(log(Offset)), data = trichoptera)
standard_error(myPLN)
}
\seealso{
\code{\link[=vcov.PLNfit]{vcov.PLNfit()}} for the complete variance covariance estimation of the coefficient
}
