% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIPLN.R
\name{ZIPLN}
\alias{ZIPLN}
\title{Zero Inflated Poisson lognormal model}
\usage{
ZIPLN(formula, data, subset, rho = 0, lambda = 0, control = list())
}
\arguments{
\item{formula}{an object of class "formula": a symbolic description of the model to be fitted. See details}

\item{data}{an optional data frame, list or environment (or object coercible by as.data.frame to a data frame) containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which PLN is called.}

\item{subset}{an optional vector specifying a subset of observations to be used in the fitting process.}

\item{rho}{a positive number controlling the level of sparsity in the inverse covariance matrix. Default to 0 (no sparsity).}

\item{lambda}{a positive number controlling the level of sparsity in the regression parameters of the PLN component. Default to 0 (no sparsity).}

\item{control}{a list for controlling the optimization. See details.}
}
\value{
an R6 object with class \code{\link{ZIPLNfit}}
}
\description{
Fit the multivariate Zero Inflated Poisson lognormal model with a variational algorithm. Use the (g)lm syntax for model specification (covariates, offsets, subset).
}
\details{
The parameter \code{control} is a list controlling the optimization with the following entries:
\itemize{
\item covariance a character setting the model for the covariance matrix: either "full", "diagonal", "spherical" or "sparse". Default is "full", automatically set to sparse if rho > 0.
\item "trace" integer for verbosity.
\item "maxiter_out" control outer optimization (VEM) maximum number of iterations in the variational E-M algorithm. Default is 100.
\item "ftol_out" control outer optimization (VEM) stop when an full V-EM iteration  changes the ELBO by less than ftol_out. Default is 1e-4.
\item "ftol_rel" control inner optimization (VE step) stop when an optimization step changes in the objective function by less than ftol multiplied by the absolute value of the parameter. Default is 1e-6 when n < p, 1e-8 otherwise.
\item "ftol_abs" control inner optimization (VE step) stop when an optimization step changes the objective function by less than ftol multiplied by the absolute value of the parameter. Default is 0
\item "xtol_rel" control inner optimization (VE step) stop when an optimization step changes every parameters by less than xtol multiplied by the absolute value of the parameter. Default is 1e-4
\item "xtol_abs" control inner optimization (VE step) stop when an optimization step changes every parameters by less than xtol multiplied by the absolute value of the parameter. Default is 0
\item "maxeval" control inner optimization (VE step) stop when the number of iteration exceeds maxeval. Default is 1000
\item "maxtime" control inner optimization (VE step) stop when the optimization time (in seconds) exceeds maxtime. Default is -1 (no restriction)
\item "algorithm" control inner optimization (VE step): the optimization method used by NLOPT among LD type, i.e. "CCSAQ", "MMA", "LBFGS", "VAR1", "VAR2". See NLOPT documentation for further details. Default is "CCSAQ".
}
}
\examples{
data(scRNA)
# data subsample: only 100 random cell and the 50 most varying transcript
scRNA        <- scRNA[sample.int(nrow(scRNA), 100), ]
scRNA$counts <- scRNA$counts[, 1:50]
myPLN_full   <- ZIPLN(counts ~ 1 + cell_line + offset(log(total_counts)) | 1, data = scRNA)
myPLN_sparse <- ZIPLN(counts ~ 1 + offset(log(total_counts)), rho = .5, data = scRNA)
myPLN_full$criteria    # better BIC with sparse version
myPLN_sparse$criteria
}
\seealso{
The class \code{\link{ZIPLNfit}}
}
