% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZIPLNfit-class.R
\name{ZIPLNfit}
\alias{ZIPLNfit}
\title{An R6 Class to represent a ZIPLNfit}
\description{
The function \code{\link[=ZIPLN]{ZIPLN()}} fit a model which is an instance of a object with class \code{\link{ZIPLNfit}}.

This class comes with a set of R6 methods, some of them being useful for the user and exported as S3 methods.
See the documentation for \code{\link[=coef]{coef()}}, \code{\link[=sigma]{sigma()}}, \code{\link[=predict]{predict()}}.

Fields are accessed via active binding and cannot be changed by the user.
}
\details{
The parameter \code{control} is a list controlling the optimization with the following entries:
\itemize{
\item covariance a character setting the model for the covariance matrix: either "full", "diagonal", "spherical" or "sparse". Default is "full", automatically set to sparse if rho > 0.
\item "trace" integer for verbosity.
\item "maxiter_out" control outer optimization (VEM) maximum number of iterations in the variational E-M algorithm. Default is 100.
\item "ftol_out" control outer optimization (VEM) stop when an full V-EM iteration  changes the ELBO by less than ftol_out. Default is 1e-4.
\item "ftol_rel" control inner optimization (VE step) stop when an optimization step changes in the objective function by less than ftol multiplied by the absolute value of the parameter. Default is 1e-6 when n < p, 1e-8 otherwise.
\item "ftol_abs" control inner optimization (VE step) stop when an optimization step changes the objective function by less than ftol multiplied by the absolute value of the parameter. Default is 0
\item "xtol_rel" control inner optimization (VE step) stop when an optimization step changes every parameters by less than xtol multiplied by the absolute value of the parameter. Default is 1e-4
\item "xtol_abs" control inner optimization (VE step) stop when an optimization step changes every parameters by less than xtol multiplied by the absolute value of the parameter. Default is 0
\item "maxeval" control inner optimization (VE step) stop when the number of iteration exceeds maxeval. Default is 1000
\item "maxtime" control inner optimization (VE step) stop when the optimization time (in seconds) exceeds maxtime. Default is -1 (no restriction)
\item "algorithm" control inner optimization (VE step): the optimization method used by NLOPT among LD type, i.e. "CCSAQ", "MMA", "LBFGS", "VAR1", "VAR2". See NLOPT documentation for further details. Default is "CCSAQ".
}
}
\examples{
\dontrun{
data(scRNA)
# data subsample: only 100 random cell and the 50 most varying transcript
subset <- sample.int(nrow(scRNA), 100)
myPLN  <- ZIPLN(counts[, 1:50] ~ 1 + offset(log(total_counts)), subset = subset, data = scRNA)
}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{n}}{number of samples}

\item{\code{p}}{number of species}

\item{\code{d}}{number of covariates}

\item{\code{latent}}{a matrix: values of the latent vector (Z in the model)}

\item{\code{latent_pos}}{a matrix: values of the latent position vector (Z) without covariates effects or offset}

\item{\code{model_par}}{a list with the matrices of parameters found in the model (B, Sigma, plus some others depending on the variant)}

\item{\code{var_par}}{a list with two matrices, M and S2, which are the estimated parameters in the variational approximation}

\item{\code{fitted}}{a matrix: fitted values of the observations (A in the model)}

\item{\code{nb_param}}{number of parameters in the current PLN model}

\item{\code{vcov_model}}{character: the model used for the covariance (either "spherical", "diagonal" or "full")}

\item{\code{optim_par}}{a list with parameters useful for monitoring the optimization}

\item{\code{loglik}}{(weighted) variational lower bound of the loglikelihood}

\item{\code{loglik_vec}}{element-wise variational lower bound of the loglikelihood}

\item{\code{BIC}}{variational lower bound of the BIC}

\item{\code{entropy}}{Entropy of the variational distribution}

\item{\code{entropy_ZI}}{Entropy of the variational distribution}

\item{\code{entropy_PLN}}{Entropy of the Gaussian variational distribution in the PLN component}

\item{\code{ICL}}{variational lower bound of the ICL}

\item{\code{criteria}}{a vector with loglik, BIC, ICL and number of parameters}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ZIPLNfit-update}{\code{ZIPLNfit$update()}}
\item \href{#method-ZIPLNfit-new}{\code{ZIPLNfit$new()}}
\item \href{#method-ZIPLNfit-optimize}{\code{ZIPLNfit$optimize()}}
\item \href{#method-ZIPLNfit-show}{\code{ZIPLNfit$show()}}
\item \href{#method-ZIPLNfit-print}{\code{ZIPLNfit$print()}}
\item \href{#method-ZIPLNfit-clone}{\code{ZIPLNfit$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZIPLNfit-update"></a>}}
\if{latex}{\out{\hypertarget{method-ZIPLNfit-update}{}}}
\subsection{Method \code{update()}}{
Update a \code{\link{ZIPLNfit}} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZIPLNfit$update(
  B = NA,
  B0 = NA,
  Pi = NA,
  Omega = NA,
  Sigma = NA,
  M = NA,
  S = NA,
  R = NA,
  Ji = NA,
  Z = NA,
  A = NA,
  monitoring = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{B}}{matrix of regression parameters in the Poisson lognormal component}

\item{\code{B0}}{matrix of regression parameters in the zero inflated component}

\item{\code{Pi}}{Zero inflated probability parameter (either scalar, row-vector, col-vector of matrix)}

\item{\code{Omega}}{precision matrix of the latent variables}

\item{\code{Sigma}}{covariance matrix of the latent variables}

\item{\code{M}}{matrix of mean vectors for the variational approximation}

\item{\code{S}}{matrix of variance parameters for the variational approximation}

\item{\code{R}}{matrix of probabilities for the variational approximation}

\item{\code{Ji}}{vector of variational lower bounds of the log-likelihoods (one value per sample)}

\item{\code{Z}}{matrix of latent vectors (includes covariates and offset effects)}

\item{\code{A}}{matrix of fitted values}

\item{\code{monitoring}}{a list with optimization monitoring quantities}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Update the current \code{\link{ZIPLNfit}} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZIPLNfit-new"></a>}}
\if{latex}{\out{\hypertarget{method-ZIPLNfit-new}{}}}
\subsection{Method \code{new()}}{
Initialize a \code{\link{ZIPLNfit}} model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZIPLNfit$new(responses, covariates, offsets, formula, xlevels, control)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{responses}}{the matrix of responses (called Y in the model). Will usually be extracted from the corresponding field in PLNfamily-class}

\item{\code{covariates}}{design matrix (called X in the model). Will usually be extracted from the corresponding field in PLNfamily-class}

\item{\code{offsets}}{offset matrix (called O in the model). Will usually be extracted from the corresponding field in PLNfamily-class}

\item{\code{formula}}{model formula used for fitting, extracted from the formula in the upper-level call}

\item{\code{xlevels}}{named listed of factor levels included in the models, extracted from the formula in the upper-level call and used for predictions.}

\item{\code{control}}{a list for controlling the optimization. See details.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZIPLNfit-optimize"></a>}}
\if{latex}{\out{\hypertarget{method-ZIPLNfit-optimize}{}}}
\subsection{Method \code{optimize()}}{
Call to the Cpp optimizer and update of the relevant fields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZIPLNfit$optimize(responses, covariates, offsets, control)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{responses}}{the matrix of responses (called Y in the model). Will usually be extracted from the corresponding field in PLNfamily-class}

\item{\code{covariates}}{design matrix (called X in the model). Will usually be extracted from the corresponding field in PLNfamily-class}

\item{\code{offsets}}{offset matrix (called O in the model). Will usually be extracted from the corresponding field in PLNfamily-class}

\item{\code{control}}{a list for controlling the optimization. See details.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZIPLNfit-show"></a>}}
\if{latex}{\out{\hypertarget{method-ZIPLNfit-show}{}}}
\subsection{Method \code{show()}}{
User friendly print method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZIPLNfit$show(
  model = paste("A multivariate Zero Inflated Poisson Lognormal fit with",
    private$covariance, "covariance model.\\n")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{First line of the print output}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZIPLNfit-print"></a>}}
\if{latex}{\out{\hypertarget{method-ZIPLNfit-print}{}}}
\subsection{Method \code{print()}}{
User friendly print method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZIPLNfit$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ZIPLNfit-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ZIPLNfit-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ZIPLNfit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
